% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCrom15Columns.R
\name{addCrom15Columns}
\alias{addCrom15Columns}
\title{Add cromatine information}
\usage{
addCrom15Columns(data)
}
\arguments{
\item{data}{dataframe used to add columns}
}
\value{
Original dataframe with cromatime related data
}
\description{
Add columns with crom15 information
}
