% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDhsColumns.R
\name{addDhsColumns}
\alias{addDhsColumns}
\title{Add dhs information}
\usage{
addDhsColumns(data)
}
\arguments{
\item{data}{dataframe used to add dhs columns}
}
\value{
Original dataframe with dhs related data
}
\description{
Add columns with dhs information
}
