% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGroupColumns.R
\name{addGroupColumns}
\alias{addGroupColumns}
\title{Add group columns from annotation data}
\usage{
addGroupColumns(data)
}
\arguments{
\item{data}{dataframe used to create columns}
}
\value{
Original dataframe with UCSC_RefGene_Group field data split in multiple column (TSS200, TSS1500, 5'UTR, 1stExon, Body, 3'UTR) with TRUE/FALSE
}
\description{
Add group columns from UCSC_RefGene_Group annotated column. UCSC_RefGene_Group column must exists on inpu dataframe
}
