% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPhantomSummary.R
\name{addPhantomSummary}
\alias{addPhantomSummary}
\title{Add phantom summary}
\usage{
addPhantomSummary(data)
}
\arguments{
\item{data}{dataframe used to create columns}
}
\value{
Original dataframe with Phantom_S field data split in multiple column (TSS200, TSS1500, 5'UTR, 1stExon, Body, 3'UTR) with TRUE/FALSE
}
\description{
Add summary from 'Phantom' annotated column in data
}
