% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_data.R
\name{adjust_data}
\alias{adjust_data}
\title{Adjust p-value by Bonferroni or FDR}
\usage{
adjust_data(cohort, cpval, bn = TRUE, fdr = TRUE, filename = NULL, N = 0)
}
\arguments{
\item{cohort}{Dataframe with CpGs related data}

\item{cpval}{String with p-value column name}

\item{bn}{Boolean if TRUE appends bonferroni adjustment column}

\item{fdr}{Boolean if TRUE appends FDR adjustment column}

\item{filename}{Optional, filename to write results, if NULL no results are writed in a file}

\item{N}{Optional, informative value to write to a file with other descriptive data related to significative p-values.}
}
\value{
Data frame with original data with bonferroni and fdr adjusted data sorted by p-value
}
\description{
Update data with Bonferroni and FDR adjustment
}
