% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_immune_response.R
\name{assess_immune_response}
\alias{assess_immune_response}
\title{Evaluate the predictive performance of the computed immune response score}
\usage{
assess_immune_response(
  predictions_immune_response = NULL,
  patient_response = NULL,
  RNA_tpm,
  TMB_values,
  easier_with_TMB = c("none", "weighted_average", "penalized_score"),
  weight_penalty = NULL,
  verbose = TRUE
)
}
\arguments{
\item{predictions_immune_response}{list containing the predictions for each
quantitative descriptor and for each task.}

\item{patient_response}{character vector with two factors
(Non-responders = NR, Responders = R).}

\item{RNA_tpm}{numeric matrix of patients' gene expression data as
tpm values.}

\item{TMB_values}{numeric vector containing patients' tumor mutational burden
(TMB) values.}

\item{easier_with_TMB}{character string indicating which approach should be
used to integrate easier with TMB.}

\item{weight_penalty}{integer value from 0 to 1, which is used to define the
weight or penalty for combining easier and TMB scores based on a weighted
average or penalized score, in order to derive a score of patient's likelihood
of immune response. The default value is 0.5.}

\item{verbose}{logical flag indicating whether to display messages about
the process.}
}
\value{
If easier_with_TMB is set to FALSE, figures two figures evaluating the
predictions are returned. If easier_with_TMB is set to TRUE, an scatterplot is
returned displaying an integrated approach that uses both immune response and
tumor mutational burden (TMB) to predict patients' response.
}
\description{
This function generates a roc curve plot and a barplot showing the average
(across tasks) area under the ROC curve (AUC) values for each quantitative
descriptor on the patients' response provided. An ensemble model is built
based on the average of the individual models.
Additionally, the average of the gold standard scores is used for comparison.
}
\examples{
# using a SummarizedExperiment object
library(SummarizedExperiment)
# Using example exemplary dataset (Mariathasan et al., Nature, 2018)
# from easierData. Original processed data is available from
# IMvigor210CoreBiologies package.
library("easierData")

dataset_mariathasan <- easierData::get_Mariathasan2018_PDL1_treatment()
RNA_tpm <- assays(dataset_mariathasan)[["tpm"]]
RNA_counts <- assays(dataset_mariathasan)[["counts"]]
cancer_type <- metadata(dataset_mariathasan)[["cancertype"]]

# Select a subset of patients to reduce vignette building time.
pat_subset <- c("SAM76a431ba6ce1", "SAMd3bd67996035", "SAMd3601288319e",
"SAMba1a34b5a060", "SAM18a4dabbc557")
RNA_counts <- RNA_counts[, colnames(RNA_counts) \%in\% pat_subset]
RNA_tpm <- RNA_tpm[, colnames(RNA_tpm) \%in\% pat_subset]

# Computation of cell fractions  (Finotello et al., Genome Med, 2019)
cell_fractions <- compute_cell_fractions(RNA_tpm = RNA_tpm)

# Computation of pathway scores (Holland et al., BBAGRM, 2019;
# Schubert et al., Nat Commun, 2018)
pathway_activities <- compute_pathway_activity(
    RNA_counts = RNA_counts,
    remove_sig_genes_immune_response = TRUE
)

# Computation of TF activity (Garcia-Alonso et al., Genome Res, 2019)
tf_activities <- compute_TF_activity(
    RNA_tpm = RNA_tpm
)

# Computation of ligand-receptor pair weights
lrpair_weights <- compute_LR_pairs(
    RNA_tpm = RNA_tpm,
    cancer_type = "pancan"
)

# Computation of cell-cell interaction scores
ccpair_scores <- compute_CC_pairs(
    lrpairs = lrpair_weights,
    cancer_type = "pancan"
)

# Predict patients' immune response
predictions <- predict_immune_response(
    pathways = pathway_activities,
    immunecells = cell_fractions,
    tfs = tf_activities,
    lrpairs = lrpair_weights,
    ccpairs = ccpair_scores,
    cancer_type = cancer_type,
    verbose = TRUE
)

# retrieve clinical response
patient_ICBresponse <- colData(dataset_mariathasan)[["BOR"]]
names(patient_ICBresponse) <- colData(dataset_mariathasan)[["pat_id"]]

# retrieve TMB
TMB <- colData(dataset_mariathasan)[["TMB"]]
names(TMB) <- colData(dataset_mariathasan)[["pat_id"]]

patient_ICBresponse <- patient_ICBresponse[names(patient_ICBresponse) \%in\% pat_subset]
TMB <- TMB[names(TMB) \%in\% pat_subset]

# Assess patient-specific likelihood of response to ICB therapy
output_eval_with_resp <- assess_immune_response(
    predictions_immune_response = predictions,
    patient_response = patient_ICBresponse,
    RNA_tpm = RNA_tpm,
    TMB_values = TMB,
    easier_with_TMB = "weighted_average",
)
}
