% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_z_score.R
\name{calc_z_score}
\alias{calc_z_score}
\title{Z-score normalization}
\usage{
calc_z_score(X, mean, sd)
}
\arguments{
\item{X}{numeric matrix.}

\item{mean}{numeric vector with mean values.}

\item{sd}{numeric vector with sd values.}
}
\value{
Numeric matrix with values as z-scores.
}
\description{
Performs z-score normalization on a numeric matrix per column.
The user can specify mean and sd values to be used to calculate z-score values,
otherwise the mean and sd is calculated based on the input matrix.
}
\examples{
# use example dataset from IMvigor210CoreBiologies package (Mariathasan et al., Nature, 2018)
data("dataset_mariathasan")
gene_tpm <- dataset_mariathasan@tpm

# apply z-score normalization
tpm_zscore <- calc_z_score(t(gene_tpm))
}
