% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_TMB.R
\name{categorize_TMB}
\alias{categorize_TMB}
\title{Convert tumor mutational burden (TMB) to categorical.}
\usage{
categorize_TMB(TMB, thresholds = NULL)
}
\arguments{
\item{TMB}{A numeric vector with tumor mutational burden values.}

\item{thresholds}{A numeric vector to specify thresholds to be used. Default thresholds are low (<100), moderate (100-400) and high TMB (>400).}
}
\value{
A numeric vector assigning each sample a class from 1 to 3.
}
\description{
\code{categorize_TMB} encodes tumor mutational variable (TMB) numerical variable to categorical.
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
rm(cds)

# retrieve Tumor Mutational Burden
TumorMutationalBurden <- clinical_data[, "FMOne mutation burden per MB"]
names(TumorMutationalBurden) <- rownames(clinical_data)

# Convert TMB continous values into categories
TMB_cat <- categorize_TMB(TMB = TumorMutationalBurden)
}
