% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromatin_enrichment.R
\name{Chromatin_enrichment}
\alias{Chromatin_enrichment}
\title{Get enrichment with Chromatin State and classify hypo i hypermethylation states}
\usage{
Chromatin_enrichment(
  data,
  out,
  filename,
  cpgcol,
  bn = FALSE,
  fdr = NA,
  pval = NA,
  all = FALSE
)
}
\arguments{
\item{data}{array or dataframe with CpGs to perform enrichment. If not beta field found then hypo and hypermethylation field is added}

\item{out}{string, Path where the results should be saved}

\item{filename}{string, File name where the results should be saved, to this name the suffix is added depending on data to enrich
\itemize{
   \item {_chrom15}{if we are enriching a CpG vector}
   \item {_chrom15_BN}{if we are enriching CpGs that meet the bonferroni condition}
   \item {_chrom15_FDR}{if we are enriching CpGs that meet the FDR condition}
   \item {_chrom15_PVAL}{if we are enriching CpGs that meet the p-value condition}
}}

\item{cpgcol}{numeric or string. Column index or column name with CpGs id}

\item{bn}{boolean. optional. If data is a dataframe with bonferroni adjustment, makes enrichment with CpGs that pass Bonferroni}

\item{fdr}{numeric optiona. If data is a dataframe with FDR adjustment and fdr!=NA or NULL, makes enrichment with CpGs with fdr lower than indicated value}

\item{pval}{numeric optional. If data is a dataframe with p-value and pval!=NA or NULL, makes enrichment with CpGs with pval lower than indicated value}

\item{all}{boolean, optional. enrich all CpGs}
}
\value{
A list with resulth enriched data.
}
\description{
Get enrichment with Chromatin States Database (chrom15). This function gets Chromatin states and creates a field meth_level for hypo and Hypermethylation states if beta field is found, if not, only enrichment is performed.
}
