% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_NA_from_data.R
\name{clean_NA_from_data}
\alias{clean_NA_from_data}
\title{Remove NA from data}
\usage{
clean_NA_from_data(data)
}
\arguments{
\item{data}{dataframe. Data frame with na data}
}
\value{
data frame without NAs
}
\description{
Remove rows with NA data and return only complete cases.
}
