% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_slash}
\alias{collapse_slash}
\title{Collapse a character vector to a slash-delimited string.}
\usage{
collapse_slash(vector, forward = TRUE, terminal_slash = TRUE)
}
\arguments{
\item{vector}{The character vector to collapse (2 or more elements).}

\item{forward}{Logical indicating whether a forward slash (TRUE) should be
used or a doubled-backslash (FALSE)}

\item{terminal_slash}{Logical indicating whether the string should be
terminated by a slash.}
}
\value{
A single element character vector containing all the elements of
\code{vector} separated by slashes.
}
\description{
Collapse a character vector to a slash-delimited string.
}
\examples{
collapse_slash(c("www.github.com", "seaaan", "easieR"))
collapse_slash(c("~", "Documents", "Important",
   "passwords-in-plain-text.txt"), terminal_slash = FALSE)
collapse_slash(c("backslashes", "are", "doubled"), forward = FALSE)
}

