% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_to_sentence}
\alias{collapse_to_sentence}
\title{Collapse a character vector to a list suitable for an English sentence.}
\usage{
collapse_to_sentence(vector, conjunction = "and", oxford_comma = TRUE,
  terminal_period = FALSE)
}
\arguments{
\item{vector}{The character vector to collapse (2 or more elements).}

\item{conjunction}{The word to use as a conjunction (no spaces).}

\item{oxford_comma}{Logical indicating whether the oxford comma should be
used.}

\item{terminal_period}{Logical indicating whether a period should be
appended.}
}
\value{
A single element character vector containing all the elements of
\code{vector} separated by commas and conjoined by conjunction.
}
\description{
Collapse a character vector to a list suitable for an English sentence.
}
\examples{
collapse_to_sentence(c("eats", "shoots", "leaves"), oxford_comma = FALSE)
collapse_to_sentence(c("my way", "the highway"), conjunction = "or")
collapse_to_sentence(1:10, conjunction = "or", terminal_period = TRUE)
}

