% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_IPS.R
\name{compute.IPS}
\alias{compute.IPS}
\title{Compute immunophenoscore}
\usage{
compute.IPS(RNA_tpm)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=Immunophenoscore
}
\description{
\code{compute_IPS} computes the immunophenoscore using source code provided by
original publication (Charoentong et al., 2017).
}
\examples{
# TODOTODO
}
