% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_signatures.R
\name{compute_Ayers_expIS}
\alias{compute_Ayers_expIS}
\title{Compute Expanded Immune signature}
\usage{
compute_Ayers_expIS(matches, RNA.tpm)
}
\arguments{
\item{matches}{TODOTODO}

\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=Expanded Immune signature score
}
\description{
\code{compute_Ayers_expIS} computes Expanded Immune signature score as the arithmetic mean of genes included
in the Expanded Immune signature (Ayers et al., JCI, 2017)
}
\examples{
# TODOTODO
}
