% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CC_pairs_grouped.R
\name{compute_CC_pairs_grouped}
\alias{compute_CC_pairs_grouped}
\title{compute_CC_pairs_grouped}
\usage{
compute_CC_pairs_grouped(lrpairs, cancertype = "pancan")
}
\arguments{
\item{lrpairs}{Ligand-receptor pairs weights matrix}

\item{cancertype}{string character}
}
\value{
A list with the following elements:
\describe{
\item{score}{cell-cell interaction scores matrix with rows=samples and columns=cells}
}
}
\description{
Compute CC pairs grouped from transcriptomics data.
}
\details{
\code{compute_CC_pairs_grouped} computes CC pairs (considering cell groups instead
of individual cell types) from tpm, using the null model for CC interaction
computed on TCGA data.
}
\examples{
# TODOTODO
}
