% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cell_fractions.R
\name{compute_cell_fractions}
\alias{compute_cell_fractions}
\title{Compute immune cell fractions from gene expression using quanTIseq}
\usage{
compute_cell_fractions(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{A data.frame containing TPM values with HGNC symbols in rows and samples in columns.}

\item{verbose}{A logical value indicating whether to display messages about the number of immune cell
signature genes found in the gene expression data provided.}
}
\value{
A matrix of normalized enrichment scores with samples in rows and cell types in columns.
}
\description{
This function estimates cell fractions from bulk RNA-seq data (tpm)
using quanTIseq method from (Finotello et al., Genome Med, 2019).
}
\examples{
# Example: Mariathasan cohort (Mariathasan et al., Nature, 2018)
if (!requireNamespace("BiocManager", quietly = TRUE))
 install.packages("BiocManager")

BiocManager::install(c("biomaRt",
 "circlize",
 "ComplexHeatmap",
 "corrplot",
 "DESeq2",
 "dplyr",
 "DT",
 "edgeR",
 "ggplot2",
 "limma",
 "lsmeans",
 "reshape2",
 "spatstat",
 "survival",
 "plyr"))

install.packages("Downloads/IMvigor210CoreBiologies_1.0.0.tar.gz", repos = NULL)
library(IMvigor210CoreBiologies)

data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

# Computation of cell fractions
cell_fractions <- compute_cell_fractions(RNA_tpm = gene_tpm)
head(cell_fractions)
}
