% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_signatures.R
\name{compute_CYT}
\alias{compute_CYT}
\title{Compute cytolytic activity score}
\usage{
compute_CYT(matches, RNA.tpm)
}
\arguments{
\item{matches}{TODOTODO}

\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=cytolytic activity score
}
\description{
\code{compute_CYT} computes cytolytic activity score as the geometric mean of immune cytolytic genes
(Rooney et al., 2015).
}
\examples{
# TODOTODO
}
