% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Davoli_IS.R
\name{compute_Davoli_IS}
\alias{compute_Davoli_IS}
\title{Compute Davoli immune signature}
\usage{
compute_Davoli_IS(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=Davoli immune signature
}
\description{
\code{compute_Davoli_IS} computes Davoli immune signature as the arithmetic mean of cytotoxic
immune infiltrate signature genes, after rank normalization (Davoli et al., 2017).
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

# Compute davoli immune signature (Davoli et al., Science 2017)
Davoli_IS <- compute_Davoli_IS(RNA_tpm= gene_tpm)
}
