% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_signatures.R
\name{compute_gene_signatures}
\alias{compute_gene_signatures}
\title{Compute gold standards}
\usage{
compute_gene_signatures(RNA.tpm, selected_signatures)
}
\arguments{
\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}

\item{selected_signatures}{TODOTODO}
}
\value{
TODOTODO
}
\description{
\code{computation_gold_standards} computes the scores for the gold standards required by the user
}
\examples{
# TODOTODO
}
