% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_signatures.R
\name{compute_ICB_genes}
\alias{compute_ICB_genes}
\title{Compute the expression of the immune checkpoints genes}
\usage{
compute_ICB_genes(genes, RNA.tpm)
}
\arguments{
\item{genes}{TODOTODO}

\item{RNA.tpm}{numeric matrix with data}
}
\value{
Data.frame with the expression of the immune checkpoint genes
}
\description{
\code{computation_ICB_genes} computes the scores for the immune checkpoint genes.
}
\examples{
# TODOTODO
}
