% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_signatures.R
\name{compute_IFNy}
\alias{compute_IFNy}
\title{Compute IFNy signature score}
\usage{
compute_IFNy(matches, RNA.tpm)
}
\arguments{
\item{matches}{TODOTODO}

\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=IFNy signature score
}
\description{
\code{compute_IFNy} computes IFNy signature score as the arithmetic mean of genes included
in the IFN-γ signature (Ayers et al., JCI, 2017)
}
\examples{
# TODOTODO
}
