% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_IMPRES.R
\name{compute_IMPRES}
\alias{compute_IMPRES}
\title{Compute Immuno-Predictive Score (IMPRES)}
\usage{
compute_IMPRES(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=IMPRES score
}
\description{
\code{compute_IMPRES} computes IMPRES score by applying logical comparison of
checkpoint gene pairs (Auslander et al., 2018).
}
\examples{
# Example: Mariathasan cohort (Mariathasan et al., Nature, 2018)
if (!requireNamespace("BiocManager", quietly = TRUE))
 install.packages("BiocManager")

BiocManager::install(c("biomaRt",
 "circlize",
 "ComplexHeatmap",
 "corrplot",
 "DESeq2",
 "dplyr",
 "DT",
 "edgeR",
 "ggplot2",
 "limma",
 "lsmeans",
 "reshape2",
 "spatstat",
 "survival",
 "plyr"))

install.packages("Downloads/IMvigor210CoreBiologies_1.0.0.tar.gz", repos = NULL)
library(IMvigor210CoreBiologies)

data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

IMPRES <- compute_IMPRES(RNA_tpm = gene_tpm)
head(IMPRES)
}
