% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_IPS.R
\name{compute_IPS}
\alias{compute_IPS}
\title{Compute immunophenoscore}
\usage{
compute_IPS(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=Immunophenoscore
}
\description{
\code{compute_IPS} computes the immunophenoscore using source code provided by
original publication (Charoentong et al., 2017).
}
\examples{
# Example: Riaz
data("Riaz_data")

IPS <- compute_IPS(RNA_tpm = Riaz_data$tpm_RNAseq)
head(IPS)
}
