% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_LR_pairs.R
\name{compute_LR_pairs}
\alias{compute_LR_pairs}
\title{Compute ligand-receptor pair weights from gene expression}
\usage{
compute_LR_pairs(
  RNA_tpm,
  remove_genes_ICB_proxies = FALSE,
  cancer_type = "pancan",
  verbose = TRUE
)
}
\arguments{
\item{RNA_tpm}{A data.frame containing TPM values with HGNC symbols in rows and samples in columns.}

\item{remove_genes_ICB_proxies}{a logical value indicating whether to remove signature genes involved
in the derivation of hallmarks of immune response.}

\item{cancer_type}{A string detailing the cancer type whose ligand-receptor pairs network will be used.
A pan-cancer network is selected by default, whose network represents the union of all
ligand-receptor pairs present across the 18 cancer types studied in (Lapuente-Santana et al., bioRxiv, 2021).}

\item{verbose}{A logical value indicating whether to display messages about the number of ligand-receptor genes found in the gene expression data provided.}
}
\value{
A matrix of weights with samples in rows and ligand-receptor pairs in columns.
}
\description{
This function quantifies ligand-receptor interactions in the tumor microenvironment from
gene expression in TPM from bulk RNA-seq data (Lapuente-Santana et al., bioRxiv, 2021), using prior knowledge
coming from ligand-receptor pair annotations from the database of (Ramilowski et al., Nat Commun, 2015).
Each ligand-receptor weight is defined as the minimum of the log2(TPM+1) expression of the ligand and the receptor.
}
\examples{
# Example: Mariathasan cohort (Mariathasan et al., Nature, 2018)
if (!requireNamespace("BiocManager", quietly = TRUE))
 install.packages("BiocManager")

BiocManager::install(c("biomaRt",
 "circlize",
 "ComplexHeatmap",
 "corrplot",
 "DESeq2",
 "dplyr",
 "DT",
 "edgeR",
 "ggplot2",
 "limma",
 "lsmeans",
 "reshape2",
 "spatstat",
 "survival",
 "plyr"))

install.packages("Downloads/IMvigor210CoreBiologies_1.0.0.tar.gz", repos = NULL)
library(IMvigor210CoreBiologies)

data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

# Computation of ligand-receptor pair weights
lrpair_weights <- compute_LR_pairs(
  RNA_tpm = gene_tpm,
  remove_genes_ICB_proxies = FALSE,
  cancer_type = "pancan")
lrpair_weights[1:5, 1:5]
}
