% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pathways_scores.R
\name{compute_pathways_scores}
\alias{compute_pathways_scores}
\title{Compute pathway activity from gene expression using PROGENy}
\usage{
compute_pathways_scores(
  RNA_counts,
  remove_genes_ICB_proxies = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{RNA_counts}{A data.frame containing raw counts values with HGNC symbols in rows and samples in columns.}

\item{remove_genes_ICB_proxies}{A logical value indicating whether to remove signature genes involved
in the derivation of hallmarks of immune response.}

\item{verbose}{A logical value indicating whether to display messages about the number of pathway signature
genes found in the gene expression data provided.}
}
\value{
A matrix with samples in rows and pathways in columns.
}
\description{
This function infers pathway activity from gene expression in raw counts from bulk RNA-seq data
using PROGENy method from (Holland et al., BBAGRM, 2019; Schubert et al., Nat Commun, 2018).
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_count <- mariathasan_data$counts
rm(cds)

# Computation of pathway scores (Holland et al., BBAGRM, 2019; Schubert et al., Nat Commun, 2018)
pathway_activity <- compute_pathways_scores(
  RNA_counts = gene_count,
  remove_genes_ICB_proxies = TRUE)
}
