% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Roh_IS.R
\name{compute_Roh_IS}
\alias{compute_Roh_IS}
\title{Compute Roh immune score}
\usage{
compute_Roh_IS(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=Roh immune score
}
\description{
\code{compute_Roh_IS} computes Roh immune score as the geometric-mean of immune
score genes (Roh et al., 2017).
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

# compute Roh immune score (Roh et al., Sci. Transl. Med., 2017)
Roh_IS <- compute_Roh_IS(RNA_tpm = gene_tpm)
}
