% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_scores_immune_response.R
\name{compute_scores_immune_response}
\alias{compute_scores_immune_response}
\title{Compute published transcriptomics-based scores
of hallmarks of anti-cancer immune response}
\usage{
compute_scores_immune_response(
  RNA_tpm = NULL,
  selected_scores = c("CYT", "Roh_IS", "chemokines", "Davoli_IS", "IFNy",
    "Ayers_expIS", "Tcell_inflamed", "RIR", "TLS"),
  verbose = TRUE
)
}
\arguments{
\item{RNA_tpm}{data.frame containing TPM values with
HGNC symbols in rows and samples in columns.}

\item{selected_scores}{character string with names of
scores of immune response to be computed. Default
scores are computed for: "CYT", "Roh_IS", "chemokines",
"Davoli_IS", "IFNy", "Ayers_expIS", "Tcell_inflamed",
"RIR", "TLS".}

\item{verbose}{logical variable indicating whether to
display informative messages.}
}
\value{
A numeric matrix with samples in rows and gold
standard scores in columns.
}
\description{
Calculates the scores of immune response as
indicated by the user.
}
\examples{
# using a SummarizedExperiment object
library(SummarizedExperiment)
# Using example exemplary dataset (Mariathasan et al., Nature, 2018)
# from easierData. Original processed data is available from
# IMvigor210CoreBiologies package.
library("easierData")

dataset_mariathasan <- easierData::get_Mariathasan2018_PDL1_treatment()
RNA_tpm <- assays(dataset_mariathasan)[["tpm"]]
cancer_type <- metadata(dataset_mariathasan)$cancertype

# Select a subset of patients to reduce vignette building time.
pat_subset <- c(
    "SAM76a431ba6ce1", "SAMd3bd67996035", "SAMd3601288319e",
    "SAMba1a34b5a060", "SAM18a4dabbc557"
)
RNA_tpm <- RNA_tpm[, colnames(RNA_tpm) \%in\% pat_subset]

# Computation of different hallmarks of anti-cancer immune responses
hallmarks_of_immune_response <- c(
    "CYT", "Roh_IS", "chemokines", "Davoli_IS", "IFNy",
    "Ayers_expIS", "Tcell_inflamed", "RIR", "TLS"
)
scores_immune_response <- compute_scores_immune_response(
    RNA_tpm = RNA_tpm,
    selected_scores = hallmarks_of_immune_response
)
}
