% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_GWAMA_files.R
\name{create_GWAMA_files}
\alias{create_GWAMA_files}
\title{Creates GWAMA formatted files}
\usage{
create_GWAMA_files(qcpath, file, gwmodelpath, N, llowCpGs = NULL)
}
\arguments{
\item{qcpath}{string. Path with Quality Control results (_QCData.txt files)}

\item{gwmodelpath}{string. Path to store gwama meta-analisis files, all gwama files related to the same meta-analysis must be under the same path}

\item{N}{number. Number of samples in model}

\item{llowCpGs}{Optiona, numeric array with all CpGs with low representation in models.}

\item{filename}{string. File name to formatt to use with GWAMA}
}
\value{
files with model data formatted for GWAMA
}
\description{
Creates files formatted with GWAMA specifications to perform a meta-analysis with GWAMA software
}
