% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crom15}
\alias{crom15}
\title{Chromatine states data}
\format{
A data frame with 482421 CpGs and 21 variables related to Chromatine states
}
\usage{
data(crom15)
}
\description{
Chromatine states data
}
\keyword{datasets}
