% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_CpGs.R
\name{descriptives_CpGs}
\alias{descriptives_CpGs}
\title{CpGs descriptive data}
\usage{
descriptives_CpGs(
  cohort,
  columns,
  filename = NULL,
  nsamples = NULL,
  before = TRUE
)
}
\arguments{
\item{cohort}{Dataframe with CpGs data}

\item{columns}{Integer vector with column position that we are interested in}

\item{filename}{Optional, filename to write results, if NULL no results are writed in a file}

\item{nsamples}{numeric, Optional, samples in cohort}

\item{before}{bolean, Optional, if true, prints text "Before QC" , else, prints "After QC"}
}
\value{
list with descriptive values
\itemize{
  \item{nCpGs}{Number of CpGs in data}
  \item{BETA}{Descriptive data for beta value}
  \item{SE}{Descriptive data for  standard Error}
  \item{P_VAL}{Descriptive data for p-value}
}
}
\description{
Get CpG descriptive data and optionally, writes a file with descriptive data from CpGs
}
