% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easier-package.R
\docType{package}
\name{easier-package}
\alias{easier}
\alias{easier-package}
\title{easier: Estimate Systems Immune Response from RNA-seq data}
\description{
This package provides a workflow for the use of EaSIeR tool, developed to assess patients' likelihood to respond to ICB therapies providing just the patients' RNA-seq data as input. We integrate RNA-seq data with different types of prior knowledge to extract quantitative descriptors of the tumor microenvironment from several points of view, including composition of the immune repertoire, and activity of intra- and extra-cellular communications. Then, we use multi-task machine learning trained in TCGA data to identify how these descriptors can simultaneously predict several state-of-the-art hallmarks of anti-cancer immune response. In this way we derive cancer-specific models and identify cancer-specific systems biomarkers of immune response. These biomarkers have been experimentally validated in the literature and the performance of EaSIeR predictions has been validated using independent datasets form four different cancer types with patients treated with anti-PD1 or anti-PDL1 therapy.
}
\author{
\strong{Maintainer}: Oscar Lapuente-Santana \email{o.lapuente.santana@tue.nl} (\href{https://orcid.org/0000-0003-1995-8393}{ORCID})

Authors:
\itemize{
  \item Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})
  \item Arsenij Ustjanzew \email{arsenij.ustjanzew@uni-mainz.de} (\href{https://orcid.org/0000-0002-1014-4521}{ORCID})
  \item Francesca Finotello \email{francesca.finotello@i-med.ac.at} (\href{https://orcid.org/0000-0003-0712-4658}{ORCID})
  \item Federica Eduati \email{f.eduati@tue.nl} (\href{https://orcid.org/0000-0002-7822-3867}{ORCID})
}

}
\keyword{internal}
