% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eQTM}
\alias{eQTM}
\title{eQTM filter}
\format{
A data frame with 104716 CpGs and 4 variables
\describe{
  \item{CpG}{CpG id}
  \item{p.value}{p-value}
  \item{sigPair}{is a significant Pair}
  \item{TC_gene}{Gene related to CpG}
}
}
\usage{
data(eQTM_autosome_unadj_filtered.cells)
}
\description{
Data for eQTM enrichment
}
\keyword{datasets}
