% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_CpGs.R
\name{exclude_CpGs}
\alias{exclude_CpGs}
\title{Exclude CpGs}
\usage{
exclude_CpGs(
  cohort,
  cpgid,
  exclude,
  ethnic,
  artype = "450K",
  filename = NULL,
  fileresume = NULL
)
}
\arguments{
\item{cohort}{Dataframe with data}

\item{cpgid}{string with the column name that contains the CpGs identifiators.}

\item{ethnic}{. Ethnicity, possible values : '' (for none ethnic filter), EUR, SAS, AMR, GWD, YRI, TSI, IBS, CHS, PUR, JPT, GIH, CH_B, STU, ITU, LWK, KHV, FIN, ESN, CEU, PJL, AC_B, CLM, CDX, GBR, BE_B, PEL, MSL, MXL, ASW or GLOBAL.}

\item{artype}{default value '450K'. Array type Illumina 450K or Illumina EPIC arrays, possible values are '450K' or 'EPIC' respectively.}

\item{filename}{Optional, filename to write excluded cpgs and related information, if NULL no data are writed in a file}

\item{fileresume}{Optional, filename to write descriptives for removed cpgs, if NULL no data are writed in a file}
}
\value{
dataframe without the CpGs that meet the conditions
}
\description{
Filter 450K or EPIC methylation array with previously defined conditions (selected conditionsin config.r). In order to minimize data size we merge only the selected ethnic column
}
