% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_biomarkers.R
\name{explore_biomarkers}
\alias{explore_biomarkers}
\title{Examination of possible relevant biomarkers in cancer dataset}
\usage{
explore_biomarkers(
  pathways = NULL,
  immunecells = NULL,
  tfs = NULL,
  lrpairs = NULL,
  ccpairs = NULL,
  cancer_type,
  real_patient_response,
  output_file_path,
  TMB_values,
  verbose = TRUE
)
}
\arguments{
\item{pathways}{A numeric matrix (rows = samples; columns = pathways).}

\item{immunecells}{A numeric matrix (rows = samples; columns = cell types).}

\item{tfs}{A numeric matrix (rows = samples; columns = transcription factors).}

\item{lrpairs}{A numeric matrix (rows = samples; columns = ligand-receptor pairs).}

\item{ccpairs}{A numeric matrix (rows = samples; columns = cell-cell pairs).}

\item{cancer_type}{A character string indicating which cancer-specific model should be used to compute the predictions.}

\item{real_patient_response}{A character vector with two factors (Non-responders = NR, Responders = R).}

\item{output_file_path}{A character string pointing to a directory to save the plots returned by the function.}

\item{verbose}{A logical flag indicating whether to display messages about the process.}
}
\value{
boxplot with features distribution
}
\description{
\code{explore_biomarkers} provides an overview of the mechanistic signatures features
distribution in the dataset (separating responders and non-responders if known)
together with information on the feature contribution to the model building.
}
\details{
Plot original features distribution together with the feature importance in
the model
}
\examples{
# TODOTODO
}
