% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_keys.R
\name{extract_keys}
\alias{extract_keys}
\title{Extract reference keys}
\usage{
extract_keys(path)
}
\arguments{
\item{path}{The path of the \code{.Rmd} file}
}
\value{
A \code{data frame} with all reference keys
}
\description{
It extracts reference keys (\emph{e.g.} bibkeys) from a \code{Rmd} file
}
\author{
Antonio J Pérez-Luque (\email{ajpelu@gmail.com})
}
