% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filter_450K}
\alias{filter_450K}
\title{450K Illumina filters}
\format{
A data frame with 485577 rows and 50 variables
}
\usage{
data(filter_450K)
}
\description{
A dataset containing the 450K Illumina methylation array probe filtering conditions
}
\references{

}
\keyword{datasets}
