% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filter_EPIC}
\alias{filter_EPIC}
\title{EPIC Illumina filters}
\format{
A data frame with 865918 rows and 50 variables
}
\usage{
data(filter_EPIC)
}
\description{
A dataset containing the EPIC Illumina methylation array probe filtering conditions
}
\references{

}
\keyword{datasets}
