% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCpGs.R
\name{filterCpGs}
\alias{filterCpGs}
\title{Filter all CpGs}
\usage{
filterCpGs(cpgs, tofilter, filter)
}
\arguments{
\item{cpgs}{string vector with CpGs}

\item{tofilter}{vector with data to apply filter}

\item{filter}{vector with filter conditions}
}
\value{
string vector with CpGs not in filter criteria
}
\description{
Filter all CpGs attending to multiple criteria values
}
