% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLowRepresentedCpGsinCohort.R
\name{filterLowRepresentedCpGsinCohort}
\alias{filterLowRepresentedCpGsinCohort}
\title{Call Rate filter}
\usage{
filterLowRepresentedCpGsinCohort(data, colname, pcmissing, Ntotal, fileresume)
}
\arguments{
\item{data}{dataframe with data to be filtered}

\item{colname}{string, column name that contains number of samples for each CpG}

\item{pcmissing}{numeric, máximum percent of missing samples allowed}

\item{Ntotal}{numeric, Number of samples in cohort}
}
\value{
filtered dataframe if colname exists
}
\description{
filters out all CpGs with missing call rates exceeding the provided value
}
