% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlapValues.R
\name{findOverlapValues}
\alias{findOverlapValues}
\title{Find Overlaps between two genomics ranges}
\usage{
findOverlapValues(selected, complete, outputdir = ".", outputfile = NULL)
}
\arguments{
\item{selected}{Genomic Ranges Variable that we are interested to enrich with data obtained from 'comlete' genomic ranges}

\item{complete}{Genomic Ranges with Fetal Placenta data 15 or 18 or other genomic ranges that overlaps with selected genomic ranges}
}
\value{
A list with the Overlap results
\itemize{
 \item{"ranges"}{A subset by overlaps}
 \item{"hits"}{Coincident data between selected and complete}
 \item{"shits"}{Subject hits}
 \item{"qhits"}{Query hits}
 \item{"values"}{Coincident names associated to each position}
}
}
\description{
Find Overlaps between two genomics ranges and returns the coincident values. Used to search overlaps between CpGs (selected data) and Fetal Placenta (complete data)
}
