% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlapValues.R
\name{findOverlapValues}
\alias{findOverlapValues}
\title{Find Overlaps between two genomics ranges}
\usage{
findOverlapValues(selected, complete)
}
\arguments{
\item{selected}{numerical. Variable to take in to account as significative variable, could be FDR, p-value,...}

\item{complete}{dataframe with all relative positions to island values to perform regression}
}
\value{
A list with the Overlap results
\itemize{
 \item{"ranges"}{A subset by overlaps}
 \item{"hits"}{Log-likelihood of linkage model}
 \item{"values"}{Coincident names}
}
}
\description{
Find Overlaps between two genomics ranges and returns the coincident values. Used to search overlaps between CpGs (selected data) and Fetal Placenta (complete data)
}
