% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats.R
\name{format_p}
\alias{format_p}
\title{Nicely display p-values at a specified number of digits.}
\usage{
format_p(p, digits)
}
\arguments{
\item{p}{A numeric vector of p-values, all between 0 and 1.}

\item{digits}{The number of digits after the decimal point to display.}
}
\value{
A character vector of formatted p-values.
}
\description{
P-values will be displayed in one of two ways:
\itemize{
   \item Like 0.01 if the rounded p-value is large enough to fit in
         \code{digits} with at least one non-zero digit.
   \item In scientific notation (e.g. 1.23e-03) otherwise. The displayed
         value will have one more digit than in \code{digits} to account
         for the digit before the decimal place.
}
}
\examples{
format_p(p = c(0.01, 0.001, 0.009, 0.5), digits = 2)
}

