% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hapmap_file.R
\name{generate_hapmap_file}
\alias{generate_hapmap_file}
\title{Create hapmap file for GWAMA}
\usage{
generate_hapmap_file(artype, outputfile = "hapmap.map")
}
\arguments{
\item{artype}{string Illumina arrat type}

\item{outputfile}{string. Optional, relative or complete path with a file name for a hapmap, if not provided hapmap file is created in current directory as hapmap.map}

\item{gwama.dir}{string. Route to GWAMA binary}
}
\value{
Disk file with hapmap
}
\description{
Create map file for GWAMA, needed in Manhattan plots. The hapmap correspondos to Illumina data and depends on array type.
}
