% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotation_unlisted_CpGs.R
\name{get_annotation_unlisted_CpGs}
\alias{get_annotation_unlisted_CpGs}
\title{Get annotation from unlisted CpGs}
\usage{
get_annotation_unlisted_CpGs(singifCpGs, artype)
}
\arguments{
\item{singifCpGs}{vector with a list of CpGs names}

\item{artype}{string with artype to obtain the unsignificative CpGs}
}
\value{

}
\description{
Get annotation (Illumina) from unlisted CpGs attending to Illumuna array type (450K or EPIC)
}
