% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations.R
\name{get_annotattions}
\alias{get_annotattions}
\title{Annotate a List of CpGs with Illumina 450K or EPIC data}
\usage{
get_annotattions(CpGs, artype, filename, outdir)
}
\arguments{
\item{CpGs}{string. Path with results from GWAMA}

\item{artype}{string, Illumina array type, 450K or EPIC}

\item{filename}{string vector with file name where annotations}

\item{outdir}{string vector. Route to output folder to store results}
}
\value{
dataframe with annotated CpGs
}
\description{
Get CpG annotations from Illumina 450K or EPIC array type
}
