% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_consensusPdb_OverRepresentation.R
\name{get_consensusPdb_OverRepresentation}
\alias{get_consensusPdb_OverRepresentation}
\title{Get ConsensusPath Over Representation Enrichment}
\usage{
get_consensusPdb_OverRepresentation(
  acFSet,
  entityType,
  accNumbers,
  accType,
  outputdir = ".",
  outputfile = NULL
)
}
\arguments{
\item{entityType}{Integer vector with column position that we are interested in}

\item{accNumbers}{Optional, filename to write results, if NULL no results are writed in a file}

\item{accType}{bolean, Optional, if true, prints text "Before QC" , else, prints "After QC"}

\item{outputfile, }{filename to write results, if NULL no results are writed in a file}
}
\value{
list with enrichment
}
\description{
Get enrichment from consensusPathDb with over representation analysis
}
