% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptives_ChromatinStates.R
\name{get_descriptives_ChromatineStates}
\alias{get_descriptives_ChromatineStates}
\title{Get descriptives Chromatine States}
\usage{
get_descriptives_ChromatineStates(
  data,
  positions,
  criteria,
  namecriteria,
  outputdir = ".",
  outputfile = NULL
)
}
\arguments{
\item{data}{string. CpG}

\item{positions}{array string. with colnames containing chromatin state positions}

\item{criteria}{boolean. Boolean with criteria like Bonferroni, FDR,...}

\item{namecriteria}{string, criteria name for plot and to add to descriptives}

\item{outputdir}{string. Output path to store file with results, by default results are
written in current dir}

\item{outputfile}{string. File name to store results if no name is provided results are
not written.The suffix "_DescChromStates.txt" and "_PlotChromStates.pdf" are added
respectively for descriptive file and plots.}
}
\value{
File with descriptive and plots from CpG Annotation
}
\description{
Get counts, proportions and plots for Chromatine States CpGs attending to any criteria as Bonferroni, FDR
}
