% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptives_GenePosition.R
\name{get_descriptives_GenePosition}
\alias{get_descriptives_GenePosition}
\title{Get descriptives gene position}
\usage{
get_descriptives_GenePosition(
  position,
  criteria,
  namecriteria,
  outputdir = ".",
  outputfile = NULL
)
}
\arguments{
\item{position}{string. CpG Gene position (Body, TSS1500, TSS200, 3'UTR, 1stExon, 5'UTR )}

\item{criteria}{boolean. Boolean with criteria like Bonferroni, FDR,...}

\item{namecriteria}{string, criteria name for plot and to add to descriptives}

\item{outputdir}{string. Output path to store file with results, by default results are written in current dir}

\item{outputfile}{string. File name to store results if no name is provided results are not written.The suffix "_DescGene.txt" and "_PlotGene.pdf" are added respectively for descriptive file and plots.}
}
\value{
File with descriptive and plots from CpG Annotation
}
\description{
Get counts, proportions and plots for gene position attending to any criteria as Bonferroni, FDR
}
