% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptives_ImprintingCarreras.R
\name{get_descriptives_ImprintingCarreras}
\alias{get_descriptives_ImprintingCarreras}
\title{Get descriptives Carreras-Gallo IRC}
\usage{
get_descriptives_ImprintingCarreras(
  position,
  criteria,
  namecriteria,
  outputdir = ".",
  outputfile = NULL
)
}
\arguments{
\item{position}{string. IRC}

\item{criteria}{boolean. Boolean with criteria like Bonferroni, FDR,...}

\item{namecriteria}{string, criteria name for plot and to add to descriptives}

\item{outputdir}{string. Output path to store file with results, by default results
are written in current dir}

\item{outputfile}{string. File name to store results if no name is provided results
are not written.The suffix "_DescCarrerasIRC.txt" and "_PlotCarrerasIRC.pdf" are added
respectively for descriptive file and plots.}
}
\value{
File with descriptive and plots from CpG Annotation
}
\description{
Get counts, proportions and plots for Carreras-Gallo IRCs attending to any criteria as
Bonferroni, FDR
}
