% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptives_postGWAMA.R
\name{get_descriptives_postGWAMA}
\alias{get_descriptives_postGWAMA}
\title{Get descriptives after GWAMA}
\usage{
get_descriptives_postGWAMA(resdir, analyzedata, modelfiles, metaname, artype)
}
\arguments{
\item{resdir}{string. Path with results from GWAMA}

\item{analyzedata}{string vector. Route to analized data in order to get descriptives}

\item{modelfiles}{string vector with models used to perform meta-analysis}

\item{metaname}{string. Meta-analisis name. only used in titles}

\item{artype}{string, Illumina array type, 450K or EPIC}
}
\value{
File with descriptive and plots from GWAMA results. Results are stored in the same file as GWAMA results
}
\description{
Get descriptives after perform GWAMA meta-analysis
}
