% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptives_RelativetoIsland.R
\name{get_descriptives_RelativetoIsland}
\alias{get_descriptives_RelativetoIsland}
\title{Get descriptives relatives to Island}
\usage{
get_descriptives_RelativetoIsland(
  position,
  criteria,
  namecriteria,
  outputdir = ".",
  outputfile = NULL
)
}
\arguments{
\item{position}{string. CpG relative position to Island (Island, N_Shelf, N_Shore, OpenSea, S_Shelf, S_Shore )}

\item{criteria}{boolean. Boolean with criteria like Bonferroni, FDR,...}

\item{namecriteria}{string, criteria name for plot and to add to descriptives}

\item{outputdir}{string. Output path to store file with results, by default results are written in current dir}

\item{outputfile}{string. File name to store results if no name is provided results are not written.The suffix "_DescIslands.txt" and "_PlotIslands.pdf" are added respectively for descriptive file and plots.}
}
\value{
File with descriptive and plots from CpG Annotation
}
\description{
Get counts, proportions and plots for relative position to Island CpGs attending to any criteria as Bonferroni, FDR
}
