% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lambda.R
\name{get_lambda}
\alias{get_lambda}
\title{Get lambda}
\usage{
get_lambda(cohort, cpval)
}
\arguments{
\item{cohort}{Dataframe with CpGs related data}

\item{cpval}{String with p-value column name}
}
\value{
lambda
}
\description{
Get lambda
}
