% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_low_presence_CpGs.R
\name{get_low_presence_CpGs}
\alias{get_low_presence_CpGs}
\title{Get CpGs with low presence}
\usage{
get_low_presence_CpGs(outputfiles, pcentMissing)
}
\arguments{
\item{outputfiles}{filename with results from GWAMA}

\item{pcentMissing}{per cent of missing data to be excluded}
}
\value{
array with low CpGs ids detected based on pcentMissing (variable from configuration file config_postQR)
}
\description{
Returns CpGs with low presence detected after first GWAS Meta-Analysis
}
