% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OE_bulk.R
\name{get_OE_bulk}
\alias{get_OE_bulk}
\title{Calculate overall expression (OE) of the immune resistance program}
\usage{
get_OE_bulk(
  r,
  gene_sign = NULL,
  num_rounds = 1000,
  full_flag = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{r}{A list containing a numeric matrix with bulk RNA-Seq data (tpm values) and a character string with the available gene names.}

\item{gene_sign}{A list containing different character strings associated with subsets of the resistance program.}

\item{num_rounds}{An integer value related to}

\item{full_flag}{A logical flag}

\item{verbose}{A logical flag indicating whether to display messages about the process.}
}
\value{
Random score
}
\description{
\code{get_OE_bulk} obtained from literature to calculate Immune resistance program
(Jerby-Arnon et al., 2018)
}
\examples{
# TODOTODO
}
