% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_semi_random_OE.R
\name{get_semi_random_OE}
\alias{get_semi_random_OE}
\title{Calculate random scores}
\usage{
get_semi_random_OE(
  r,
  genes_dist_q,
  b_sign,
  num_rounds = 1000,
  full_flag = FALSE
)
}
\arguments{
\item{r}{list TODOTODO - more info here as well}

\item{genes_dist_q}{integer}

\item{b_sign}{boolean}

\item{num_rounds}{integer}

\item{full_flag}{boolean}
}
\value{
Random score
}
\description{
\code{get_semi_random_OE} obtained from literature to calculate Immune resistance
program  (Jerby-Arnon et al., 2018) TODOTODO - out of sync?
}
\examples{
# TODOTODO
}
