% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllChromStateRegressions.R
\name{getAllChromStateRegressions}
\alias{getAllChromStateRegressions}
\title{Significative variable Regression for all cromatin state}
\usage{
getAllChromStateRegressions(
  significative,
  chromstate,
  outputdir = ".",
  outputfile = NULL
)
}
\arguments{
\item{significative}{numerical. Variable to take in to account as significative variable, could be FDR, p-value,...}

\item{outputdir}{string. Output path to store file with results, by default results are written in current dir}

\item{outputfile}{string. File name to store results if no name is provided results are not written.The suffix "RegressionFDR_States" is added to file name provided.}

\item{chromstates}{dataframe with all Cromatin state values to perform regression}
}
\value{

}
\description{
Get regression by cromatin state taking in to account a significative variable parameter like FDR values. This function gets all chromatin state regressions and optionally store results in a file
}
