% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllRelativeIslandOR.R
\name{getAllRelativeIslandOR}
\alias{getAllRelativeIslandOR}
\title{Significative variable Regression for all relative positions to island}
\usage{
getAllRelativeIslandOR(
  significative,
  position,
  outputdir = ".",
  outputfile = NULL,
  plots = TRUE
)
}
\arguments{
\item{significative}{numerical. Variable to take in to account as significative variable, could be FDR, p-value,...}

\item{position}{dataframe with all relative positions to island values to perform regression}

\item{outputdir}{optional string. Output path to store file with results, by default results are written in current dir}

\item{outputfile}{optional string. File name to store results if no name is provided results are not written.}

\item{plots}{boolean. If plot is TRUE, plot results}
}
\value{

}
\description{
Get regression by relative positions to island position taking in to account a significative variable parameter like FDR values. This function gets all chromatin state regressions and optionally store results in a file
}
