% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBinaryClassificationYesNo.R
\name{getBinaryClassificationYesNo}
\alias{getBinaryClassificationYesNo}
\title{Get Binary classification (yes / no)}
\usage{
getBinaryClassificationYesNo(value, condition = "<", limit = 0)
}
\arguments{
\item{value}{vector with values to classify as binary "yes" / "no"}

\item{condition}{Condition to apply in order to classify values, "<", ">", "<=" or ">=", by default condition is "<"}

\item{limit}{limit value to classify data taking into account "condition" param, by default limit is 0}
}
\value{
Vector with betas length with classification in Hyper and Hypo
}
\description{
Get binary classificatoin from vector values taking into account a logical operator
}
