% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChromStateOR.R
\name{getChromStateOR}
\alias{getChromStateOR}
\title{Significative variable Regression by cromatin state}
\usage{
getChromStateOR(significative, chromstate, varname)
}
\arguments{
\item{significative}{numerical. Variable to take in to account as significative variable, could be FDR, p-value,...}

\item{chromstate}{vector. Cromatin state values}

\item{varname}{string. Cromatin state name. For example : "TssA","TssAFlnk","TxFlnk","TxWk","Tx","EnhG","Enh","ZNF.Rpts","Het","TssBiv","BivFlnk","EnhBiv","ReprPC","ReprPCWk" or "Quies"}
}
\value{

}
\description{
Get regression by cromatin state taking in to account a significative variable parameter like FDR values
}
