% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichGenomicRanges.R
\name{getEnrichGenomicRanges}
\alias{getEnrichGenomicRanges}
\title{Create GRanges from data}
\usage{
getEnrichGenomicRanges(seqnames, start, end)
}
\arguments{
\item{seqnames}{string vector with seqnames}

\item{start}{numeric vector with start position}

\item{end}{numeric vector with end position}
}
\value{
Data transformed in genomic ranges with name as chr;start;end
}
\description{
Create Genomic Ranges from data
}
