% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntrezIdfromSymbol.R
\name{getEntrezIdfromSymbol}
\alias{getEntrezIdfromSymbol}
\title{Get Entrez gene ID}
\usage{
getEntrezIdfromSymbol(symbols)
}
\arguments{
\item{symbols}{array with gene symbols}
}
\value{
Vector with Entrez gene IDs
}
\description{
Get Entrez gene ID from gene Symbol
}
