% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geteQTMEnrichment.R
\name{geteQTMEnrichment}
\alias{geteQTMEnrichment}
\title{Get eQTM enrichment}
\usage{
geteQTMEnrichment(CpGs, outputdir = ".", outputfile = NULL, plots = TRUE)
}
\arguments{
\item{CpGs}{vector with a list of CpGs names}

\item{outputdir}{string. Output path to store file with results, by default results are written in current dir}

\item{outputfile}{string. File name to store results if no name is provided results are not written.The suffix "RegressionFDR_States" is added to file name provided.}

\item{plots}{boolean. If plot is TRUE, plot results}
}
\value{
gene list
}
\description{
Get Gene Ontology Enrichment analysis for the eQTM genes
}
