% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFisherTest.R
\name{getFisherTest}
\alias{getFisherTest}
\title{Get Fisher test for significative variable attending to criteria}
\usage{
getFisherTest(significative, criteria, varname)
}
\arguments{
\item{significative}{numerical. Variable to take in to account as significative variable, could be FDR, p-value,...}

\item{criteria}{vector with criteria}

\item{varname}{string. criteria Relative to Island name. For example : "Island", "OpenSea", "S_Shore", "N_Shore", "S_Shelf", "N_Shelf" or other criteria like gene position}
}
\description{
Get Fisher test by criteria taking in to account a significative variable parameter like FDR values
}
