% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHypergeometricTest.R
\name{getHypergeometricTest}
\alias{getHypergeometricTest}
\title{HyperGeometric Test for positions relative to Island}
\usage{
getHypergeometricTest(significative, criteria, varname)
}
\arguments{
\item{varname}{string. position Relative to Island name. For example : "Island", "OpenSea", "S_Shore", "N_Shore", "S_Shelf", "N_Shelf"}

\item{selected}{vector with elements that meet conditions}

\item{complete}{vector with all possible cases that can meet conditions}
}
\value{

}
\description{
Get HyperGeometric Test for positions relative to Island
}
