% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHyperHypo.R
\name{getHyperHypo}
\alias{getHyperHypo}
\title{Get Hiper and Hipo methylation data}
\usage{
getHyperHypo(betas)
}
\arguments{
\item{betas}{vector with beta values to classify as hyper and hypo methylated}
}
\value{
Vector with betas length with classification in Hyper and Hypo
}
\description{
Get columns with Hyper and Hypo Methylation information
}
