% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPMDGenomicRanges.R
\name{getPMDGenomicRanges}
\alias{getPMDGenomicRanges}
\title{Create GRanges from PMD}
\usage{
getPMDGenomicRanges(seqnames, start, end)
}
\arguments{
\item{seqnames}{string vector with seqnames}

\item{start}{numeric vector with start position}

\item{end}{numeric vector with end position}
}
\value{
Data transformed in genomic ranges with name as chr;start;end
}
\description{
Create Genomic Ranges from PMD
}
