% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniqueValues.R
\name{getUniqueValues}
\alias{getUniqueValues}
\title{Get unique values in vector}
\usage{
getUniqueValues(x)
}
\arguments{
\item{x}{Vector with strings separated by ";"}
}
\value{
Vector with unique string values
}
\description{
Get unique values from a Vector with strings separated by ";"
}
