% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hs.c2}
\alias{Hs.c2}
\title{Molecular Signatures Database (MSigDB v 7.1) - C2 curated gene sets}
\format{
A data frame with 5529 variables
}
\usage{
data(human_c2_v7p1)
}
\description{
A dataset containing Molecular Signatures Database in his current version 7.1.
}
\references{
\url{http://bioinf.wehi.edu.au/MSigDB/v7.1/}
}
\keyword{datasets}
