% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IR_Placenta}
\alias{IR_Placenta}
\title{Imprinting Regions Placenta}
\format{
A data frame with 485 regions and 6 variables
\describe{
  \item{Chr_DMR}{Chromosome}
  \item{Start_DMR}{Start position}
  \item{End_DMR}{End position}
  \item{DMR_Classification}{DMR classification, possible values are "Candidate mDMR", "Known placenta-specific mDMR", "Candidate pDMR", "Known mDMR" and "Known pDMR".}
  \item{Known_gDMR }{Known DMR associated gene }
}
}
\usage{
data(IR_Placenta)
}
\description{
Imprinting regions for placenta
}
\keyword{datasets}
