% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IRCarreras_Ev1}
\alias{IRCarreras_Ev1}
\title{Imprinting Regions Carreras-Gallo Epic V1}
\format{
A data frame with 156 regions
\describe{
  \item{ICR_id}{ICR_id}
  \item{CpG_chr}{Chromosome}
  \item{CpG_start}{Start position}
  \item{CpG_stop}{End position}
  \item{ICR_chr}{Chromosome}
  \item{ICR_start}{Start position}
  \item{ICR_stop}{End position}
  \item{CpG_Probe}{DMR classification, possible values are "Candidate mDMR", "Known placenta-specific mDMR", "Candidate pDMR", "Known mDMR" and "Known pDMR".}
  \item{ID}{Original ICR ID}
  \item{Genom_Coord}{Genomic Coordinates}
  \item{Meth_Origin}{Parental Origin of Methylation}
  \item{Near_Trasncript}{Nearest Transcript}
  \item{Dist_Near_Trasncript}{Distance to Nearest Transcript}
}
}
\usage{
data(IRCarreras_Ev1)
}
\description{
Imprinting regions for Epic Version 1
}
\keyword{datasets}
