% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missMethyl_enrichment.R
\name{missMethyl_enrichment}
\alias{missMethyl_enrichment}
\title{Get Pathways enrichment with missMethyl}
\usage{
missMethyl_enrichment(
  data,
  out,
  filename,
  artype = "450K",
  bn = FALSE,
  fdr = NA,
  pval = NA,
  all = FALSE,
  plots = FALSE
)
}
\arguments{
\item{data}{array or dataframe with CpGs to perform enrichment. Dataframe option is used to filter by Bonferroni FDR or p-value, in order to filter by this fields, function needs that Bonferroni with adjustments field  to be called "Bonferroni", the field with the FDR adjustments to be called FDR and the field with the p-value to be called p.value}

\item{out}{string, Path where the results should be saved}

\item{filename}{string, File name where the results should be saved, to this name the suffix is added depending on data to enrich
\itemize{
   \item {_mysmeth}{if we are enriching a CpG vector}
   \item {_mysmeth_BN}{if we are enriching CpGs that meet the bonferroni condition}
   \item {_mysmeth_FDR}{if we are enriching CpGs that meet the FDR condition}
   \item {_mysmeth_PVAL}{if we are enriching CpGs that meet the p-value condition}
}}

\item{artype}{string, Illumina array type, 450K or EPIC, by default array type is 450K}

\item{bn}{boolean. optional. If data is a dataframe with bonferroni adjustmnet, makes enrichment with CpGs that pass Bonferroni}

\item{fdr}{numeric optiona. If data is a dataframe with FDR adjustmnet and fdr!=NA or NULL, makes enrichment with CpGs with fdr lower than indicated value}

\item{pval}{numeric optional. If data is a dataframe with p-value and pval!=NA or NULL, makes enrichment with CpGs with pval lower than indicated value}

\item{all}{boolean, optional. enrich all CpGs}

\item{plots}{boolean, optional. generate plots with GO and KEGG data}
}
\value{
File with descriptive and plots from GWAMA results. Results are stored in the same file as GWAMA results
}
\description{
Get Pathways enrichment with missMethyl Bioconductor package. This function gets GO enrichment and KEGG enrichment from input data and save the results in a two separated files, one with GO enrichment and the other with KEGG enrichment with _GO and _KEGG sufix respectively.
}
