% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSigDB_enrichment.R
\name{MSigDB_enrichment}
\alias{MSigDB_enrichment}
\title{Get enrichment with Molecular Signatures Database (MSigDB)}
\usage{
MSigDB_enrichment(
  data,
  out,
  filename,
  artype = "450K",
  bn = FALSE,
  fdr = NA,
  pval = NA,
  all = FALSE
)
}
\arguments{
\item{data}{array or dataframe with CpGs to perform enrichment. Dataframe option is used to filter by Bonferroni FDR or p-value, in order to filter by this fields, function needs that Bonferroni with adjustments field  to be called "Bonferroni", the field with the FDR adjustments to be called "FDR" and the field with the p-value to be called "p.value"}

\item{out}{string, Path where the results should be saved}

\item{filename}{string, File name where the results should be saved, to this name the suffix is added depending on data to enrich
\itemize{
   \item {_MSigDB}{if we are enriching a CpG vector}
   \item {_MSigDB_BN}{if we are enriching CpGs that meet the bonferroni condition}
   \item {_MSigDB_FDR}{if we are enriching CpGs that meet the FDR condition}
   \item {_MSigDB_PVAL}{if we are enriching CpGs that meet the p-value condition}
}}

\item{artype}{string, Illumina array type, 450K or EPIC, by default array type is 450K}

\item{bn}{boolean. optional. If data is a dataframe with bonferroni adjustmnet, makes enrichment with CpGs that pass Bonferroni}

\item{fdr}{numeric optiona. If data is a dataframe with FDR adjustmnet and fdr!=NA or NULL, makes enrichment with CpGs with fdr lower than indicated value}

\item{pval}{numeric optional. If data is a dataframe with p-value and pval!=NA or NULL, makes enrichment with CpGs with pval lower than indicated value}

\item{all}{boolean, optional. enrich all CpGs}
}
\value{
A list with resulth enriched data. Results also are stored in a folder MSigDB inside Enrichment default path
}
\description{
Get enrichment with Molecular Signatures Database (MSigDB c2 - v7p1). This function gets MSigDB enrichment and save the results in  _mysmeth sufix. Data from :  Walter and Eliza Hall Institute website for local bioinformatic resources. (http://bioinf.wehi.edu.au/MSigDB/v7.1/)
}
