% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chromosomestate.R
\name{plot_chromosomestate}
\alias{plot_chromosomestate}
\title{Chromosome State Plot}
\usage{
plot_chromosomestate(
  data,
  outputdir = ".",
  outputfile = NULL,
  main = "",
  xlab = "",
  ...
)
}
\arguments{
\item{main}{optional, string with title}

\item{xlab}{optional, string with xlab text}

\item{x}{Dataframe with Chromosome State data}
}
\value{
distribution plot
}
\description{
Chromosome State plot
}
