% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ChromStates.R
\name{plot_ChromStates}
\alias{plot_ChromStates}
\title{Plot ChromStates}
\usage{
plot_ChromStates(
  x,
  outputdir = ".",
  outputfile = NULL,
  main = "",
  xlab = "",
  ...
)
}
\arguments{
\item{x}{Dataframe with ChromStates OR data}

\item{outputdir}{string with relative path}

\item{main}{optional, string with title}

\item{xlab}{optional, string with xlab text}
}
\value{
distribution plot
}
\description{
Plot ChromStates
}
