% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Depletion_Enrichment.R
\name{plot_Depletion_Enrichment}
\alias{plot_Depletion_Enrichment}
\title{Depletion and Enrichment Plot}
\usage{
plot_Depletion_Enrichment(
  x,
  outputdir = ".",
  outputfile = NULL,
  main = "",
  xlab = "",
  ...
)
}
\arguments{
\item{x}{Dataframe with Depletion and Enrichment data}

\item{outputdir}{string with relative path}

\item{main}{optional, string with title}

\item{xlab}{optional, string with xlab text}
}
\value{
distribution plot
}
\description{
Depletion and Enrichment Plot
}
