% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Distribution plot}
\usage{
plot_distribution(x, main = "", xlab = "", ...)
}
\arguments{
\item{x}{Data}

\item{main}{optional, string with title}

\item{xlab}{optional, string with xlab text}
}
\value{
distribution plot
}
\description{
Distribution plot
}
