% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forestplot.R
\name{plot_ForestPlot}
\alias{plot_ForestPlot}
\title{Forest plot}
\usage{
plot_ForestPlot(
  datas,
  files_meta,
  islowCpg,
  gwana_dir,
  metaname,
  files,
  outputgwama,
  nsignificatives = 30
)
}
\arguments{
\item{datas}{complete data obtained from meta-analysis after QC without annotations}

\item{files_meta}{string vector with files used in the meta-analysis}

\item{islowCpg}{string  string indicating if the execution has been carried out with all the CpGs or with the filtered CpGs, possible values : 'Normal', 'lowcpgs'}

\item{gwana_dir}{string with gwama input data path}

\item{metaname}{string with meta-analysis name}

\item{files}{string vector with files used in QC because we need data from this files to perform ForestPlots}

\item{outputgwama}{string with gwama output path}

\item{nsignificatives}{number, lowest p-values to show in forestplot, by default, lowestp=30}
}
\value{
distribution plot
}
\description{
Forest plot
}
