% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GenePosition.R
\name{plot_GenePosition}
\alias{plot_GenePosition}
\title{Gene Pisition Plots}
\usage{
plot_GenePosition(
  x,
  outputdir = ".",
  outputfile = NULL,
  main = "",
  xlab = "",
  ...
)
}
\arguments{
\item{x}{Dataframe with Gene Position data}

\item{outputdir}{string with relative path}

\item{main}{optional, string with title}

\item{xlab}{optional, string with xlab text}
}
\value{
distribution plot
}
\description{
Gene Pisition plot
}
