% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OR.R
\name{plot_OR}
\alias{plot_OR}
\title{Plot OR}
\usage{
plot_OR(x, outputdir = ".", outputfile = NULL, main = "", xlab = "", ...)
}
\arguments{
\item{x}{Dataframe with OR data}

\item{outputdir}{string with relative path}

\item{main}{optional, string with title}

\item{xlab}{optional, string with xlab text}
}
\value{
distribution plot
}
\description{
Plot OR
}
