% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_precisionp.R
\name{plot_precisionp}
\alias{plot_precisionp}
\title{Precision plot}
\usage{
plot_precisionp(x, filename, main = NULL, ...)
}
\arguments{
\item{x}{Data}

\item{filename}{optional, file name to write the plot with png format}

\item{main}{optional, string with plot title}
}
\value{
distribution plot
}
\description{
Precision plot
}
