% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_precissionp.R
\name{plot_precissionp}
\alias{plot_precissionp}
\title{Precission plot}
\usage{
plot_precissionp(x, filename, main = NULL, ...)
}
\arguments{
\item{x}{Data}

\item{filename}{optional, file name to write the plot with png format}

\item{main}{optional, string with plot title}
}
\value{
distribution plot
}
\description{
Precission plot
}
