% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RelativetoIsland.R
\name{plot_RelativetoIsland}
\alias{plot_RelativetoIsland}
\title{Relative to Island Plot}
\usage{
plot_RelativetoIsland(
  x,
  outputdir = ".",
  outputfile = NULL,
  main = "",
  xlab = "",
  ...
)
}
\arguments{
\item{x}{Dataframe with Relative to Island data}

\item{outputdir}{string with relative path}

\item{main}{optional, string with title}

\item{xlab}{optional, string with xlab text}
}
\value{
distribution plot
}
\description{
Relative to Island plot
}
