% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Summary.R
\name{plot_Summary}
\alias{plot_Summary}
\title{Plot enrichment summary}
\usage{
plot_Summary(df, xl, yl, tit, filename = NULL)
}
\arguments{
\item{df}{Dataframe with enrichment data}

\item{xl}{string with x axis title}

\item{yl}{string with y axis title}

\item{tit}{string with main title}

\item{filename}{optional string. File name to store results if no name is provided results are not written.}
}
\value{
plot
}
\description{
Plot enrichment summary from GO and KEGG
}
