% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TestEnrichDepleit_Collapsed.R
\name{plot_TestEnrichDepleit_Collapsed}
\alias{plot_TestEnrichDepleit_Collapsed}
\title{Plot collapsed tests results}
\usage{
plot_TestEnrichDepleit_Collapsed(
  x,
  outputdir = ".",
  outputfile = NULL,
  main = "",
  xlab = "",
  ...
)
}
\arguments{
\item{x}{list with datataframes to plot}

\item{outputdir}{string with relative path}

\item{main}{optional, string with title}

\item{xlab}{optional, string with xlab text}
}
\value{
distribution plot
}
\description{
Plot collapsed tests results for Enriched and Depleit
}
