% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TestResults_Collapsed_Positions.R
\name{plot_TestResults_Collapsed_Positions}
\alias{plot_TestResults_Collapsed_Positions}
\title{Plot collapsed tests results for positions}
\usage{
plot_TestResults_Collapsed_Positions(
  x,
  outputdir = ".",
  outputfile = NULL,
  main = "",
  xlab = "",
  ...
)
}
\arguments{
\item{x}{list with datataframes to plot}

\item{outputdir}{string with relative path}

\item{main}{optional, string with title}

\item{xlab}{optional, string with xlab text}
}
\value{
distribution plot
}
\description{
Plot collapsed tests results for Gene position and Relative to Island position
}
