% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venndiagram.R
\name{plot_venndiagram}
\alias{plot_venndiagram}
\title{Venn diagrams plots}
\usage{
plot_venndiagram(
  venndata,
  qcpath = ".",
  plotpath = ".",
  pattern = "*_QCData.txt",
  bn = NULL,
  fdr = NULL
)
}
\arguments{
\item{venndata}{prefixes used as a file name to store QC data}

\item{qcpath}{optional, path to store Venn Diagrams, by default current path}

\item{pattern}{optional, pattern with data to be used in Venn diagrams, by default '*_QCData.txt'}

\item{bn, }{optional, column name with bonferroni adjusted values, if NULL, no venndiagram plot for Bonferroni}

\item{fdr, }{column name with bonferroni adjusted values, if NULL, no venndiagram plot for FDR}

\item{path}{optional, path where QC data is stored, by default current path}
}
\value{
Venn diagrams
}
\description{
Plot Venn diagrams
}
