% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot}
\usage{
plot_volcano(x, cbeta, cpval, nCpGs, main = "", ...)
}
\arguments{
\item{x}{Data}

\item{cbeta}{string with beta column name}

\item{cpval}{string with p-value column name}

\item{main}{optional, string with title}

\item{xlab}{optional, string with xlab text}
}
\value{
distribution plot
}
\description{
Volcano plot
}
