% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PMD_placenta}
\alias{PMD_placenta}
\title{PMD placenta}
\format{
A data frame with 3 variables
}
\usage{
data(PMD_placenta)
}
\description{
PMD placenta
}
\references{
Schroeder, D. I. et al. The human placenta methylome. Proc. Natl. Acad. Sci. 110, 6037–6042 (2013)
}
\keyword{datasets}
