% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_with_rmtlr.R
\name{predict_with_rmtlr}
\alias{predict_with_rmtlr}
\title{Function to compute regularized multi-task linear regression predictions}
\usage{
predict_with_rmtlr(
  view_name,
  view_info,
  view_data,
  cancer_type,
  verbose = TRUE
)
}
\arguments{
\item{view_name}{A character string containing the name of the input view.}

\item{view_info}{A character string informing about the family of the input data.}

\item{view_data}{A list containing the data for each input view.}

\item{cancer_type}{A character string indicating cancer type to specify cancer-specific optimization model to be used.}

\item{verbose}{A logical flag indicating whether to display messages about the process.}
}
\value{
A list of predictions matrices, one for each tasks (rows = samples; columns = runs).
}
\description{
\code{predict_with_rmtlr} predicts patients' immune response using a cancer-specifc model learned from training.
}
\examples{
# Example: Mariathasan cohort (Mariathasan et al., Nature, 2018)
if (!requireNamespace("BiocManager", quietly = TRUE))
 install.packages("BiocManager")

BiocManager::install(c("biomaRt",
 "circlize",
 "ComplexHeatmap",
 "corrplot",
 "DESeq2",
 "dplyr",
 "DT",
 "edgeR",
 "ggplot2",
 "limma",
 "lsmeans",
 "reshape2",
 "spatstat",
 "survival",
 "plyr"))

install.packages("Downloads/IMvigor210CoreBiologies_1.0.0.tar.gz", repos = NULL)
library(IMvigor210CoreBiologies)

data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

# Computation of cell fractions
cell_fractions <- compute_cell_fractions(RNA_tpm = gene_tpm)

view_name <- "immunecells"
view_info <- c(immunecells = "gaussian")
view_data <- list(immunecells = as.data.frame(cell_fractions))

# Predict using rmtlr
prediction_view <- predict_with_rmtlr(
view_name = view_name,
view_info = view_info,
view_data = view_data,
cancer_type = "SKCM"
)
}
