% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_mariathasan.R
\name{preprocess_mariathasan}
\alias{preprocess_mariathasan}
\title{Pre-process data from mariathasan cohort}
\usage{
preprocess_mariathasan(cds, verbose = TRUE)
}
\arguments{
\item{cds}{CountDataSet object}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
A list with three data.frames:
\describe{
\item{counts}{A data.frame containing gene count data with genes as rows and samples as columns.}
\item{tpm}{A data.frame containing tpm data with genes as rows and samples as columns.}
\item{clinical}{A data.frame that includes information on certain clinical variables.}
}
}
\description{
\code{preprocess_mariathasan} prepares data from mariathasan cohort to be used along the vignette.
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
data_mariathasan <- preprocess_mariathasan(cds)
}
\references{
Mariathasan, S., Turley, S., Nickles, D. et al. TGFβ attenuates tumour response to PD-L1 blockade by contributing to exclusion of T cells.
Nature 554, 544–548 (2018). https://doi.org/10.1038/nature25501
}
