% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTableEnrich.R
\name{printTableEnrich}
\alias{printTableEnrich}
\title{Print table with enrichment data}
\usage{
printTableEnrich(x, columns = c(1:6))
}
\arguments{
\item{x}{dataframe with data to print}

\item{columns}{columns to show in table}
}
\value{
void
}
\description{
Prints table with enrichment data
}
