% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicate_CpGs.R
\name{remove_duplicate_CpGs}
\alias{remove_duplicate_CpGs}
\title{Remove duplicate CpGs}
\usage{
remove_duplicate_CpGs(cohort, colname, descfile = NULL, duplifile = NULL)
}
\arguments{
\item{cohort}{Dataframe with data}

\item{descfile}{Optional, filename to write data description if NULL no descriptive values write}

\item{duplifile}{Optional, filename to write duplicates if NULL no results stored in a file}
}
\value{
cohort data without duplicated values
}
\description{
Get a list with information related to duplicated values in data
}
