% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_na}
\alias{replace_na}
\title{Replace NA values with \code{replacement}.}
\usage{
replace_na(vector, replacement, warn = TRUE)
}
\arguments{
\item{vector}{Vector to replace NA values in.}

\item{replacement}{Value to replace NAs with (must be a single element).}

\item{warn}{Logical indicating whether to issue a warning message if no
element is replaced.}
}
\value{
\code{Vector} with all NA elements replaced by \code{replacement}.
}
\description{
Replace NA values with \code{replacement}.
}
\examples{
replace_na(c("a", "b", NA), "look at me")
}

