% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_regex}
\alias{replace_regex}
\title{Replace elements matching a regular expression with \code{replacement}.}
\usage{
replace_regex(vector, replacement, regex, warn = TRUE)
}
\arguments{
\item{vector}{Vector to replace matches in.}

\item{replacement}{Value to replace matches with (must be a single element).}

\item{regex}{A regular expression (must be a single element).}

\item{warn}{Logical indicating whether to issue a warning message if no
element is replaced.}
}
\value{
\code{Vector} with all elements matching \code{regex} replaced by
\code{replacement}.

Note that factors will be releveled with the replacement values and levels
no longer present in the factor may be lost.
}
\description{
Replace elements matching a regular expression with \code{replacement}.
}
\examples{
replace_value(c("Farmworker", "Programmer", "Farmhand", "Pole vaulter",
      "Farmer", "Dancer"), replacement = "Farmer", regex = "^Farm.")
}

