% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_value}
\alias{replace_value}
\title{Replace values identical to \code{value} with \code{replacement}.}
\usage{
replace_value(vector, value, replacement, warn = TRUE)
}
\arguments{
\item{vector}{Vector to replace matches in.}

\item{value}{Value to be replaced (must be a single element).}

\item{replacement}{Value to replace matches with (must be a single element).}

\item{warn}{Logical indicating whether to issue a warning message if no
element is replaced.}
}
\value{
\code{Vector} with all instances of \code{value} replaced by
\code{replacement}.

Note that factors will be releveled with the replacement values and levels
no longer present in the factor may be lost.
}
\description{
Replace values identical to \code{value} with \code{replacement}.
}
\examples{
replace_value(c("Lions", "Tigers", "Antelopes"), value = "Antelopes",
   replacement = "Bears")
}

