% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replace.R
\name{replace_values}
\alias{replace_values}
\title{Replace multiple unique elements with multiple unique replacements.}
\usage{
replace_values(vector, values, replacements, warn = TRUE)
}
\arguments{
\item{vector}{Vector to make replacements in.}

\item{values}{Values to be replaced (vector of same length as
\code{replacements}).}

\item{replacements}{Values to replace matches with (vector of same length as
\code{values}).}

\item{warn}{Logical indicating whether to issue a warning message if no
element is replaced.}
}
\value{
Vector with elements in \code{values} replaced with the
corresponding elements from \code{replacements}.

Every element in \code{values} will be searched for in \code{vector}. If it is
found, it will be replaced in \code{vector} with the corresponding element of
\code{replacements}. That is, if the 3rd element of \code{values} is found
anywhere in \code{vector}, it will be replaced by the 3rd element of
\code{replacements}.

Note that factors will be releveled with the replacement values and levels
no longer present in the factor may be lost.
}
\description{
Replace multiple unique elements with multiple unique replacements.
}
\examples{
replace_values(
   vector = c("WA", "WY", "OR", "NY", "NY", "WA", "FL", "OR"),
   values = c("FL", "NY", "OR", "WA", "WY"),
   replacements = c("Florida", "New York", "Oregon", "Washington", "Wyoming"))
replace_values(
   vector = c("M", "F", "F", "M", "F", "M", "F", "M", "M", "M", "F", "F"),
   values = c("F", "M"),
   replacements = c("Female", "Male"))
}

