% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictCpGstoArtype.R
\name{restrict_CpGs_to_artype}
\alias{restrict_CpGs_to_artype}
\title{Restrict CpGs to array type}
\usage{
restrict_CpGs_to_artype(CpGSites, artype)
}
\arguments{
\item{CpGSites, }{string vector with CpG sites}

\item{artype}{string, Illumina array type, 450K or EPIC}
}
\value{
dataframe with data restricted to Illymina array type /450K or EPIC)
}
\description{
Restrict CpGs to array type. In EPIC array uses only CpGs in 450K arrays
}
