% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyDuplicates.R
\name{simplifyDuplicates}
\alias{simplifyDuplicates}
\title{Get unique values in a cell}
\usage{
simplifyDuplicates(x)
}
\arguments{
\item{x}{strings with duplicated data separated by ";"}
}
\value{
string with unique values separated by ;
}
\description{
Get unique values in a cell with data separated by ;
}
