% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_norm.R
\name{standardization}
\alias{standardization}
\title{Calculation of a matrix z-score from scratch.}
\usage{
standardization(X, mean, sd)
}
\arguments{
\item{X}{numeric matrix with data}

\item{mean}{numeric vector with data}

\item{sd}{numeric vector with data}
}
\value{
numeric matrix with scaled data
}
\description{
\code{standardization} implements the z-score normalization.
}
\examples{
# TODOTODO
}
