% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_States_FP_Table.R
\name{summary_States_FP_Table}
\alias{summary_States_FP_Table}
\title{Summarized States Fetal Placenta for significant, hyper and hypo in a data frame}
\usage{
summary_States_FP_Table(
  significative,
  significativehyper,
  significativehypo,
  criteria,
  resformat = "long",
  outputdir = ".",
  outputfile = NULL,
  plot = TRUE
)
}
\arguments{
\item{significative, }{boolean vector with significant = 'yes' in significant values for BN, FDR, p-value ....}

\item{significativehyper}{boolean vector with significanthyper='yes' if CpG is significant for BN, FDR, p-value ... and CpG is hyper-methylated}

\item{significativehypo}{boolean vector with significanthypo='yes' if CpG is significant for BN, FDR, p-value ... and CpG is hyper-methylated}

\item{criteria}{string vector with values to take in to account to perform summarize, for example Chromatine States 15 Fetal Placenta}

\item{resformat}{string, if 'long' returns a extended table to use with plot else returns a compacted table}

\item{outputdir}{string. Output path to store file with results, by default results are written in current dir}

\item{outputfile}{string. File name to store results if no name is provided results are not written.The suffix "_States_FP" and the prefix "Porp" are added to file name provided.}

\item{plot}{boolean. If plot is TRUE, plot results}
}
\value{
Dataframe with summarized data (extended version)
}
\description{
Summarized States Fetal Placenta for significant, hyper and hypo in a data frame
}
