% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_duplicate_CpGs.R
\name{test_duplicate_CpGs}
\alias{test_duplicate_CpGs}
\title{Test duplicate CpGs}
\usage{
test_duplicate_CpGs(cohort, colname, duplifile = NULL)
}
\arguments{
\item{cohort}{Dataframe with data}

\item{colname}{string, column name to search duplicates}

\item{descfile}{Optional, filename to write data description if NULL no descriptive values write}
}
\value{
cohort data without duplicated values
}
\description{
Get a list with information related to duplicated values in data
}
