% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_QCData.R
\name{write_QCData}
\alias{write_QCData}
\title{Write QC results to a file}
\usage{
write_QCData(cohort, filename)
}
\arguments{
\item{cohort}{Dataframe with data}

\item{filename}{File name to write Quality Control results, func}
}
\value{
void
}
\description{
Write results from Quaity Control to a file. In filename the prefix _QCData is added in order to use this data in next steps like Meta-analysis
}
