% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EasieR.R
\name{zPlot}
\alias{zPlot}
\title{Building a function to translate a score from distribution in a z-score graph.}
\usage{
zPlot(my_x, myMean, mySd, p = 0, y)
}
\arguments{
\item{my_x}{is the score of interest from a given distribution.}

\item{myMean}{is the mean of the given distribution.}

\item{mySd}{is the standard deviation of the given distribution.}

\item{p}{default zero, it is called for plotting and calculating "less" P(x<a), "more" P(x>a), "between" P(a<x<b)}

\item{y}{second score for a "probability between x and y" problem}
}
\value{
a plot stressing the corresponding z-score of the input data.
}
\description{
This function is thought for the chapter 6 on Standard Normal distributions. It allows to rapidly draw a plot with Lattice representing a Standard normal distribution translating the inputs for a generic normal distribution. Furthermore, it allows to stress out cumulative probability distributions if necessary.
}
\details{
the package is actually under development.
}
\author{
Maurizio Mario Murino
}
\seealso{
\code{Lattice, xyplot}
}

