% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_immune_response.R
\name{assess_immune_response}
\alias{assess_immune_response}
\title{Evaluate the predictive performance of the computed immune response score}
\usage{
assess_immune_response(
  predictions_immune_response = NULL,
  patient_response = NULL,
  RNA_tpm,
  TMB_values,
  easier_with_TMB = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{predictions_immune_response}{list containing the predictions for each quantitative descriptor and for each task.}

\item{patient_response}{character vector with two factors (Non-responders = NR, Responders = R).}

\item{RNA_tpm}{numeric matrix of patients' gene expression data as tpm values.}

\item{TMB_values}{numeric vector containing patients' tumor mutational burden (TMB) values.}

\item{easier_with_TMB}{logical flag indicating whether to combine both easier and TMB scores based on a weighted
average or penalized score (a weight, penalty of 0.5 is used).}

\item{verbose}{logical flag indicating whether to display messages about the process.}
}
\value{
If easier_with_TMB is set to FALSE, two figures (roc curve and bar plots) are directly saved in the path specified
in output_file_path. If easier_with_TMB is set to TRUE, an additional plot is returned displaying an integrated approach
that uses both immune response and tumor mutational burden (TMB) to predict patients' response.
}
\description{
This function generates a roc curve plot and a barplot showing the average
(across tasks) area under the ROC curve (AUC) values for each quantitative descriptor on
the patients' response provided. An ensemble model is built based on the average of the individual models.
Additionally, the average of the gold standard scores is used for comparison.
}
\examples{
# Load exemplary dataset (Mariathasan et al., Nature, 2018) from easierData.
# Original processed data is available from IMvigor210CoreBiologies package.
library("easierData")
dataset_mariathasan <- easierData::get_Mariathasan2018_PDL1_treatment()
RNA_tpm <- dataset_mariathasan@assays@data@listData[["tpm"]]
RNA_counts <- dataset_mariathasan@assays@data@listData[["counts"]]
cancer_type <- dataset_mariathasan@metadata$cancertype

# Select a subset of patients to reduce vignette building time.
set.seed(1234)
subset <- sample(colnames(RNA_tpm), size = 30)
RNA_counts <- RNA_counts[, subset]
RNA_tpm <- RNA_tpm[, subset]

# Computation of cell fractions  (Finotello et al., Genome Med, 2019)
cell_fractions <- compute_cell_fractions(RNA_tpm = RNA_tpm)

# Computation of pathway scores (Holland et al., BBAGRM, 2019; Schubert et al., Nat Commun, 2018)
pathway_activities <- compute_pathway_activity(
  RNA_counts = RNA_counts,
  remove_sig_genes_immune_response = TRUE
)

# Computation of TF activity (Garcia-Alonso et al., Genome Res, 2019)
tf_activities <- compute_TF_activity(
  RNA_tpm = RNA_tpm
)

# Computation of ligand-receptor pair weights
lrpair_weights <- compute_LR_pairs(
  RNA_tpm = RNA_tpm,
  cancer_type = "pancan"
)

# Computation of cell-cell interaction scores
ccpair_scores <- compute_CC_pairs(
  lrpairs = lrpair_weights,
  cancer_type = "pancan"
)

# Predict patients' immune response
predictions <- predict_immune_response(
  pathways = pathway_activities,
  immunecells = cell_fractions,
  tfs = tf_activities,
  lrpairs = lrpair_weights,
  ccpairs = ccpair_scores,
  cancer_type = cancer_type,
  verbose = TRUE
)

# retrieve clinical response
patient_ICBresponse <- dataset_mariathasan@colData$BOR
names(patient_ICBresponse) <- dataset_mariathasan@colData$pat_id

# retrieve TMB
TMB <- dataset_mariathasan@colData$TMB
names(TMB) <- dataset_mariathasan@colData$pat_id

patient_ICBresponse <- patient_ICBresponse[subset]
TMB <- TMB[subset]

# Assess patient-specific likelihood of response to ICB therapy
assess_immune_response(
  predictions_immune_response = predictions,
  patient_response = patient_ICBresponse,
  RNA_tpm = RNA_tpm,
  TMB_values = TMB,
  easier_with_TMB = TRUE
)
}
