% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_z_score.R
\name{calc_z_score}
\alias{calc_z_score}
\title{Z-score normalization}
\usage{
calc_z_score(X, mean, sd)
}
\arguments{
\item{X}{A numeric matrix.}

\item{mean}{A numeric vector with mean values.}

\item{sd}{A numeric vector with sd values.}
}
\value{
A numeric matrix with values as z-scores.
}
\description{
\code{calc_z_score} applies z-score normalization on a numeric matrix per column.
The user can specify mean and sd values to be used to calculate z-score values,
otherwise the mean and sd is calculated based on the input matrix.
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)
# apply z-score normalization
tpm_zscore <- calc_z_score(t(gene_tpm))
}
