% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_z_score.R
\name{calc_z_score}
\alias{calc_z_score}
\title{Perform Z-score normalization}
\usage{
calc_z_score(X, mean, sd)
}
\arguments{
\item{X}{numeric matrix.}

\item{mean}{numeric vector with mean values.}

\item{sd}{numeric vector with standard deviation values.}
}
\value{
A numeric matrix with values as z-scores.
}
\description{
This function performs z-score normalization on a numeric matrix per column.
Mean and standard deviation values to be used to calculate z-score values, otherwise they are
calculated based on the input matrix.
}
\examples{
# Load exemplary dataset (Mariathasan et al., Nature, 2018) from ExperimentHub easierData.
# Original processed data is available from IMvigor210CoreBiologies package.
library("ExperimentHub")
eh <- ExperimentHub()
easierdata_eh <- query(eh, c("easierData"))
dataset_mariathasan <- easierdata_eh[["EH6677"]]
RNA_tpm <- dataset_mariathasan@assays@data@listData[["tpm"]]

# apply z-score normalization
tpm_zscore <- calc_z_score(t(RNA_tpm))
}
