% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_TMB.R
\name{categorize_TMB}
\alias{categorize_TMB}
\title{Convert tumor mutational burden (TMB) to categorical.}
\usage{
categorize_TMB(TMB, thresholds = NULL)
}
\arguments{
\item{TMB}{numeric vector with tumor mutational burden values.}

\item{thresholds}{numeric vector to specify thresholds to be used. Default thresholds are low (<100),
moderate (100-400) and high TMB (>400).}
}
\value{
A numeric vector assigning each sample a class from 1 to 3.
}
\description{
This function encodes tumor mutational variable (TMB) numerical variable to categorical.
}
\examples{
# Load exemplary dataset (Mariathasan et al., Nature, 2018) from easierData.
# Original processed data is available from IMvigor210CoreBiologies package.
library("easierData")
dataset_mariathasan <- easierData::get_Mariathasan2018_PDL1_treatment()
TMB <- dataset_mariathasan@colData$TMB; names(TMB) <- dataset_mariathasan@colData$pat_id

# Select a subset of patients to reduce vignette building time.
set.seed(1234)
subset <- sample(names(TMB), size = 30)
TMB <- TMB[subset]

# Convert TMB continous values into categories
TMB_cat <- categorize_TMB(TMB = TMB)
}
