% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_TMB.R
\name{categorize_TMB}
\alias{categorize_TMB}
\title{Convert tumor mutational burden (TMB) to categorical.}
\usage{
categorize_TMB(x, thresholds = NULL)
}
\arguments{
\item{thresholds}{A numeric vector to specify thresholds to be used. Default thresholds are low (<100), moderate (100-400) and high TMB (>400).}

\item{TMB}{A numeric vector with tumor mutational burden values.}
}
\value{
A numeric vector assigning each sample a class from 1 to 3.
}
\description{
\code{categorize_TMB} encodes tumor mutational variable (TMB) numerical variable to categorical.
}
\examples{
# Example: Mariathasan cohort (Mariathasan et al., Nature, 2018)
if (!requireNamespace("BiocManager", quietly = TRUE))
 install.packages("BiocManager")

BiocManager::install(c("biomaRt",
 "circlize",
 "ComplexHeatmap",
 "corrplot",
 "DESeq2",
 "dplyr",
 "DT",
 "edgeR",
 "ggplot2",
 "limma",
 "lsmeans",
 "reshape2",
 "spatstat",
 "survival",
 "plyr"))

install.packages("Downloads/IMvigor210CoreBiologies_1.0.0.tar.gz", repos = NULL)
library(IMvigor210CoreBiologies)

data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
rm(cds)

# retrieve Tumor Mutational Burden
TumorMutationalBurden <- clinical_data[, "FMOne mutation burden per MB"]
names(TumorMutationalBurden) <- rownames(clinical_data)

TMB_cat <- categorize_TMB(TumorMutationalBurden)
TMB_cat
}
