% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Ayers_expIS.R
\name{compute_Ayers_expIS}
\alias{compute_Ayers_expIS}
\title{Compute Expanded Immune signature}
\usage{
compute_Ayers_expIS(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=Expanded Immune signature score
}
\description{
\code{compute_Ayers_expIS} computes Expanded Immune signature score as the arithmetic
mean of genes included in the Expanded Immune signature (Ayers et al., JCI, 2017)
}
\examples{
# Example: Riaz
data("Riaz_data")

Ayers_expIS <- compute_Ayers_expIS(RNA_tpm= Riaz_data$tpm_RNAseq)
head(Ayers_expIS)
}
