% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Ayers_expIS.R
\name{compute_Ayers_expIS}
\alias{compute_Ayers_expIS}
\title{Compute Expanded Immune signature}
\usage{
compute_Ayers_expIS(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=Expanded Immune signature score
}
\description{
\code{compute_Ayers_expIS} computes Expanded Immune signature score as the arithmetic
mean of genes included in the Expanded Immune signature (Ayers et al., JCI, 2017)
}
\examples{
# Example: Mariathasan cohort (Mariathasan et al., Nature, 2018)
if (!requireNamespace("BiocManager", quietly = TRUE))
 install.packages("BiocManager")

BiocManager::install(c("biomaRt",
 "circlize",
 "ComplexHeatmap",
 "corrplot",
 "DESeq2",
 "dplyr",
 "DT",
 "edgeR",
 "ggplot2",
 "limma",
 "lsmeans",
 "reshape2",
 "spatstat",
 "survival",
 "plyr"))

install.packages("Downloads/IMvigor210CoreBiologies_1.0.0.tar.gz", repos = NULL)
library(IMvigor210CoreBiologies)

data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

Ayers_expIS <- compute_Ayers_expIS(gene_tpm)
head(Ayers_expIS)
}
