% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CC_pairs_grouped.R
\name{compute_CC_pairs_grouped}
\alias{compute_CC_pairs_grouped}
\title{Compute cell-cell interactions scores from gene expression}
\usage{
compute_CC_pairs_grouped(lrpairs, cancertype = "pancan", verbose = TRUE)
}
\arguments{
\item{lrpairs}{A matrix of weights log2(TPM +1) with samples in rows and ligand-receptor pairs in columns. This data is returned by compute_LR_pairs function.}

\item{cancertype}{A string detailing the cancer type whose cell-cell interaction network will be used.
A pan-cancer network is selected by default, whose network represents the union of all
ligand-receptor pairs present across the 18 cancer types studied in (Lapuente-Santana et al., bioRxiv, 2021).}

\item{verbose}{A logical value indicating whether to display informative messages about the process.}
}
\value{
A matrix of scores with samples in rows and cell-cell pairs in columns.
}
\description{
This function scores cell-cell interactions in the tumor microenvironment from
gene expression in TPM from bulk RNA-seq data (Lapuente-Santana et al., 2021), using prior knowledge
coming from ligand-receptor pair annotations from the database of (Ramilowski et al., Nat Commun, 2015).
}
\examples{
# Example: Riaz
data("Riaz_data")
lrpairs_weights <- compute_LR_pairs(
  RNA_tpm = Riaz_data$tpm_RNAseq,
  remove_genes_ICB_proxies = FALSE,
  cancertype = "pancan")

# Computation of cell-cell interaction scores
ccpair_scores <- compute_CC_pairs_grouped(
  lrpairs = lrpairs_weights,
  cancertype = "pancan")
head(ccpair_scores)
}
