% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CC_pairs_grouped.R
\name{compute_CC_pairs_grouped}
\alias{compute_CC_pairs_grouped}
\title{compute_CC_pairs_grouped}
\usage{
compute_CC_pairs_grouped(lrpairs, cancertype = "pancan", verbose = TRUE)
}
\arguments{
\item{lrpairs}{Ligand-receptor pairs weights matrix}

\item{cancertype}{string character}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
Cell-Cell interaction scores matrix with rows=samples and columns=cell-cell interactions
}
\description{
Compute CC pairs grouped from transcriptomics data.
}
\details{
\code{compute_CC_pairs_grouped} computes CC pairs (considering cell groups instead
of individual cell types) from tpm, using the null model for CC interaction
computed on TCGA data.
}
\examples{
# Example: Riaz
data("Riaz_data")
lrpairs_weights <- compute_LR_pairs(
  RNA_tpm = Riaz_data$tpm_RNAseq,
  remove_genes_ICB_proxies = FALSE,
  cancertype = "pancan")

# Computation of cell-cell interaction scores
ccpair_scores <- compute_CC_pairs_grouped(
  lrpairs = lrpairs_weights,
  cancertype = "pancan")
head(ccpair_scores)
}
