% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CCpair_score.R
\name{compute_CCpair_score}
\alias{compute_CCpair_score}
\title{compute_CCpair_score}
\usage{
compute_CCpair_score(
  celltype1,
  celltype2,
  intercell_network,
  lrpairs_binary,
  lr_frequency,
  compute_log = TRUE,
  cancertype
)
}
\arguments{
\item{celltype1}{string character with first cell type involved in the interaction}

\item{celltype2}{string character with second cell type involved in the interaction}

\item{intercell_network}{matrix with data on cell types interaction network}

\item{lrpairs_binary}{binary vector displaying LR pairs with non-zero frequency}

\item{lr_frequency}{numeric vector with LR pairs frequency across the whole TCGA database}

\item{compute_log}{boolean variable in order to take the log of the weighted score}

\item{cancertype}{string character}
}
\value{
numeric vector with weighted scores
}
\description{
\code{compute_CCpair_score} computes the score for each cell-cell pair for all the patients.
}
\examples{
# TODOTODO
}
