% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cell_fractions.R
\name{compute_cell_fractions}
\alias{compute_cell_fractions}
\title{Compute immune cell fractions}
\usage{
compute_cell_fractions(RNA_tpm)
}
\arguments{
\item{RNA_tpm}{numeric matrix of tpm values with rows=genes and columns=samples}
}
\value{
Cell fractions matrix: matrix of normalized enrichment scores with
rows=samples and columns=TFs
}
\description{
\code{compute_cell_fractions} estimates cell fractions from bulk RNA-seq data.
}
\details{
Compute cell fractions from transcriptomics data. This function computes cell
fractions from tpm RNA-seq data using quanTIseq method
}
\examples{
# Example: Riaz
data("Riaz_data")

# Computation of cell fractions
cell_fractions <- compute_cell_fractions(RNA_tpm = Riaz_data$tpm_RNAseq)
}
