% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_signatures.R
\name{compute_chemokines}
\alias{compute_chemokines}
\title{Compute chemokine score}
\usage{
compute_chemokines(matches, RNA.tpm)
}
\arguments{
\item{matches}{TODOTODO}

\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=chemokine score
}
\description{
\code{compute_chemokine} computes chemoine score as the PC1 score that results from applying PCA
to z-score expression of 12 chemokine genes (Messina et al., 2012).
}
\examples{
# TODOTODO
}
