% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_chemokines.R
\name{compute_chemokines}
\alias{compute_chemokines}
\title{Compute chemokine score}
\usage{
compute_chemokines(RNA_tpm)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=chemokine score
}
\description{
\code{compute_chemokines} computes chemokine score as the PC1 score that results from
applying PCA to z-score expression of 12 chemokine genes (Messina et al., 2012).
}
\examples{
# Example: Riaz
data("Riaz_data")

chemokines <- compute_chemokines(RNA_tpm = Riaz_data$tpm_RNAseq)
head(chemokines)
}
