% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_signatures.R
\name{compute_CYT}
\alias{compute_CYT}
\title{Compute cytolytic activity (CYT) score}
\usage{
compute_CYT(matches, RNA_tpm, verbose)
}
\arguments{
\item{matches}{numeric vector indicating the index of signature genes in \code{RNA_tpm}.}

\item{RNA_tpm}{data.frame containing TPM values with HGNC symbols in rows and samples in columns.}

\item{verbose}{logical value indicating whether to display informative messages.}
}
\value{
A numeric matrix with samples in rows and CTY score in a column.
}
\description{
This function calculates the CYT score as the geometric mean of its signature genes.
}
\examples{
# use example dataset from IMvigor210CoreBiologies package (Mariathasan et al., Nature, 2018)
data("dataset_mariathasan")
gene_tpm <- dataset_mariathasan@tpm

# Compute cytolytic activity (Rooney et al, Cell, 2015)
CYT <- compute_CYT(RNA_tpm = gene_tpm)
}
\references{
Rooney, M.S., Shukla, S.A., Wu, C.J., Getz, G., and Hacohen, N. (2015). Molecular and genetic properties
of tumors associated with local immune cytolytic activity. Cell 160, 48–61. https://doi.org/10.1016/j.cell.2014.12.033.
}
