% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CYT.R
\name{compute_CYT}
\alias{compute_CYT}
\title{Compute cytolytic activity score}
\usage{
compute_CYT(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=cytolytic activity score
}
\description{
\code{compute_CYT} computes cytolytic activity score as the geometric mean of immune
cytolytic genes (Rooney et al., 2015).
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

# Compute cytolytic activity (Rooney et al, Cell, 2015)
CYT <- compute_CYT(RNA_tpm = gene_tpm)
}
