% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CYT.R
\name{compute_CYT}
\alias{compute_CYT}
\title{Compute cytolytic activity score}
\usage{
compute_CYT(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=cytolytic activity score
}
\description{
\code{compute_CYT} computes cytolytic activity score as the geometric mean of immune
cytolytic genes (Rooney et al., 2015).
}
\examples{
# Example: Mariathasan cohort (Mariathasan et al., Nature, 2018)
if (!requireNamespace("BiocManager", quietly = TRUE))
 install.packages("BiocManager")

BiocManager::install(c("biomaRt",
 "circlize",
 "ComplexHeatmap",
 "corrplot",
 "DESeq2",
 "dplyr",
 "DT",
 "edgeR",
 "ggplot2",
 "limma",
 "lsmeans",
 "reshape2",
 "spatstat",
 "survival",
 "plyr"))

install.packages("Downloads/IMvigor210CoreBiologies_1.0.0.tar.gz", repos = NULL)
library(IMvigor210CoreBiologies)

data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

CYT <- compute_CYT(RNA_tpm = gene_tpm)
head(CYT)
}
