% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Davoli_IS.R
\name{compute_Davoli_IS}
\alias{compute_Davoli_IS}
\title{Compute Davoli immune signature}
\usage{
compute_Davoli_IS(RNA_tpm)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=Davoli immune signature
}
\description{
\code{compute_Davoli_IS} computes Davoli immune signature as the arithmetic mean of cytotoxic
immune infiltrate signature genes, after rank normalization (Davoli et al., 2017).
}
\examples{
# Example: Riaz
data("Riaz_data")

Davoli_IS <- compute_Davoli_IS(RNA_tpm= Riaz_data$tpm_RNAseq)
head(Davoli_IS)
}
