% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gold_standards.R
\name{compute_gold_standards}
\alias{compute_gold_standards}
\title{Compute published transcriptomics-based scores of hallmarks of anti-cancer immune response
(so-called gold standards)}
\usage{
compute_gold_standards(
  RNA_tpm,
  list_gold_standards = c("CYT", "Roh_IS", "chemokines", "Davoli_IS", "IFNy",
    "Ayers_expIS", "Tcell_inflamed", "RIR", "TLS")
)
}
\arguments{
\item{RNA_tpm}{data.frame containing TPM values with HGNC symbols in rows and samples in columns.}

\item{list_gold_standards}{character string of task names to be considered as gold standards for comparison.}
}
\value{
A numeric matrix with samples in rows and gold standard scores in columns.
}
\description{
The function computes the gold standard scores required by the user.
}
\examples{
# use example dataset from IMvigor210CoreBiologies package (Mariathasan et al., Nature, 2018)
data("dataset_mariathasan")
gene_tpm <- dataset_mariathasan@tpm

# Computation of different hallmarks of the immune response
tasks <- c(
  "CYT", "Roh_IS", "chemokines", "Davoli_IS", "IFNy",
  "Ayers_expIS", "Tcell_inflamed", "RIR", "TLS"
)
tasks_values <- compute_gold_standards(
  RNA_tpm = gene_tpm,
  list_gold_standards = tasks
)
}
