% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gold_standards.R
\name{compute_gold_standards}
\alias{compute_gold_standards}
\title{Compute gold standards}
\usage{
compute_gold_standards(
  RNA.tpm,
  list_gold_standards,
  cancertype,
  output_file_path
)
}
\arguments{
\item{RNA.tpm}{numeric matrix with data, as tpm values}

\item{list_gold_standards}{string with gold standards names}

\item{cancertype}{string character}

\item{output_file_path}{TODOTODO to define - maybe name it more specifically to the function it absolves?}
}
\value{
List with the scores of all the gold standards specified.
}
\description{
\code{computation_gold_standards} computes the scores for the gold standards required by the user
}
\examples{
# TODOTODO
}
