% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_IFNy.R
\name{compute_IFNy}
\alias{compute_IFNy}
\title{Compute IFNy signature score}
\usage{
compute_IFNy(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=IFNy signature score
}
\description{
\code{compute_IFNy} computes IFNy signature score as the arithmetic mean of genes
included in the IFN-γ signature (Ayers et al., JCI, 2017)
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

# Compute IFNy signature (Ayers et al., JCI, 2017)
IFNy <- compute_IFNy(RNA_tpm = gene_tpm)
}
