% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_IFNy.R
\name{compute_IFNy}
\alias{compute_IFNy}
\title{Compute IFNy signature (IFNy) score}
\usage{
compute_IFNy(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=IFNy signature score
}
\description{
Computes IFNy signature score as the average expression of its signature genes
}
\examples{
# use example dataset from IMvigor210CoreBiologies package (Mariathasan et al., Nature, 2018)
data("dataset_mariathasan")
gene_tpm <- dataset_mariathasan@tpm

# Compute IFNy signature (Ayers et al., JCI, 2017)
IFNy <- compute_IFNy(RNA_tpm = gene_tpm)
}
\references{
Ayers, M., Lunceford, J., Nebozhyn, M., Murphy, E., Loboda, A., Kaufman, D.R., Albright,
A., Cheng, J.D., Kang, S.P., Shankaran, V., et al. (2017). IFN-γ-related mRNA profile predicts clinical
response to PD-1 blockade. J. Clin. Invest. 127, 2930–2940. https://doi.org/10.1172/JCI91190.
}
