% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_IMPRES.R
\name{compute_IMPRES}
\alias{compute_IMPRES}
\title{Compute Immuno-Predictive Score (IMPRES)}
\usage{
compute_IMPRES(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=IMPRES score
}
\description{
\code{compute_IMPRES} computes IMPRES score by applying logical comparison of
checkpoint gene pairs (Auslander et al., 2018).
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

# Compute IMPRES signature (Auslander et al., Nat.Med., 2018)
IMPRES <- compute_IMPRES(RNA_tpm = gene_tpm)
}
