% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_IMPRES.R
\name{compute_IMPRES}
\alias{compute_IMPRES}
\title{Compute Immuno-Predictive Score (IMPRES)}
\usage{
compute_IMPRES(RNA_tpm)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=IMPRES score
}
\description{
\code{compute_IMPRES} computes IMPRES score by applying logical comparison of
checkpoint gene pairs (Auslander et al., 2018).
}
\examples{
# Example: Riaz
data("Riaz_data")

IMPRES <- compute_IMPRES(RNA_tpm = Riaz_data$tpm_RNAseq)
head(IMPRES)
}
