% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_signatures.R
\name{compute_IMPRES_MSI}
\alias{compute_IMPRES_MSI}
\title{Compute Immuno-Predictive Score (IMPRES)}
\usage{
compute_IMPRES_MSI(sig, len, match_F_1, match_F_2, RNA.tpm)
}
\arguments{
\item{sig}{TODOTODO}

\item{len}{TODOTODO}

\item{match_F_1}{TODOTODO}

\item{match_F_2}{TODOTODO}

\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=IMPRES score
}
\description{
\code{compute_IMPRES} computes IMPRES score by applying logical comparison of checkpoint gene pairs
(Auslander et al., 2018).
}
\examples{
# TODOTODO
}
