% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_LR_pairs.R
\name{compute_LR_pairs}
\alias{compute_LR_pairs}
\title{Compute ligand-receptor pairs}
\usage{
compute_LR_pairs(
  RNA_tpm,
  remove_genes_ICB_proxies = FALSE,
  cancertype = "pancan"
)
}
\arguments{
\item{RNA_tpm}{numeric matrix of tpm values with rows=genes and columns=samples}

\item{remove_genes_ICB_proxies}{boolean variable to remove all those genes
involved in the computation of ICB proxy's of response}

\item{cancertype}{string character}
}
\value{
Ligand-receptor pairs weights matrix in log2(tpm + 1) with rows=samples and columns=L-R pairs
}
\description{
\code{compute_LR_pairs} obtain ligand-receptor pairs weights from tpm RNA-seq data.
}
\examples{
# Example: Riaz
data("Riaz_data")

# Computation of ligand-receptor pair weights
lrpairs_weights <- compute_LR_pairs(
  RNA_tpm = Riaz_data$tpm_RNAseq,
  remove_genes_ICB_proxies = FALSE,
  cancertype = "pancan")
lrpairs_weights[1:5, 1:5]
}
