% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_LR_pairs.R
\name{compute_LR_pairs}
\alias{compute_LR_pairs}
\title{Compute ligand-receptor pair weights from TPM bulk gene expression}
\usage{
compute_LR_pairs(
  RNA_tpm,
  remove_genes_ICB_proxies = FALSE,
  cancer_type = "pancan",
  verbose = TRUE
)
}
\arguments{
\item{RNA_tpm}{A data.frame containing TPM values with HGNC symbols in rows and samples in columns.}

\item{remove_genes_ICB_proxies}{a logical value indicating whether to remove signature genes involved
in the derivation of hallmarks of immune response.}

\item{cancer_type}{A string detailing the cancer type whose ligand-receptor pairs network will be used.
A pan-cancer network is selected by default, whose network represents the union of all
ligand-receptor pairs present across the 18 cancer types studied in (Lapuente-Santana et al., bioRxiv, 2021).}

\item{verbose}{A logical value indicating whether to display messages about the number of ligand-receptor
genes found in the gene expression data provided.}
}
\value{
A matrix of weights with samples in rows and ligand-receptor pairs in columns.
}
\description{
This function quantifies ligand-receptor interactions in the tumor microenvironment from
TPM bulk gene expression (Lapuente-Santana et al., bioRxiv, 2021), using prior knowledge
coming from ligand-receptor pair annotations from the database of (Ramilowski et al., Nat Commun, 2015).
Each ligand-receptor weight is defined as the minimum of the log2(TPM+1) expression of
the ligand and the receptor.
}
\examples{
# use example dataset from IMvigor210CoreBiologies package (Mariathasan et al., Nature, 2018)
data("dataset_mariathasan")
gene_tpm <- dataset_mariathasan@tpm

# Computation of ligand-receptor pair weights
lrpair_weights <- compute_LR_pairs(
  RNA_tpm = gene_tpm,
  remove_genes_ICB_proxies = FALSE,
  cancer_type = "pancan"
)
lrpair_weights[1:5, 1:5]
}
