% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_MSI.R
\name{compute_MSI}
\alias{compute_MSI}
\title{Compute MSI score}
\usage{
compute_MSI(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=MSI score
}
\description{
\code{compute_MSI} computes MSI score by applying logical comparison of MSI-related
gene pairs (Fu et al., 2019).
}
\examples{
# Example: Riaz
data("Riaz_data")

MSI <- compute_MSI(RNA_tpm = Riaz_data$tpm_RNAseq)
head(MSI)
}
