% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_MSI.R
\name{compute_MSI}
\alias{compute_MSI}
\title{Compute microsatellite instability status (MSI) score}
\usage{
compute_MSI(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=MSI score
}
\description{
Computes MSI score by logical comparison of MSI-related gene pairs expression
}
\examples{
# use example dataset from IMvigor210CoreBiologies package (Mariathasan et al., Nature, 2018)
data("dataset_mariathasan")
gene_tpm <- dataset_mariathasan@tpm

# Compute Microsatellite instability status (Fu et al., BMC Genomics, 2019 )
MSI <- compute_MSI(RNA_tpm = gene_tpm)
}
\references{
Fu, Y., Qi, L., Guo, W., Jin, L., Song, K., You, T., Zhang, S., Gu, Y., Zhao, W.,
and Guo, Z. (2019). A qualitative transcriptional signature for predicting microsatellite instability
status of right-sided Colon Cancer. BMC Genomics 20, 769. https://doi.org/10.1186/s12864-019-6129-8.
}
