% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pathways_scores.R
\name{compute_pathways_scores}
\alias{compute_pathways_scores}
\title{Compute pathways score}
\usage{
compute_pathways_scores(
  RNA_counts,
  remove_genes_ICB_proxies = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{RNA_counts}{numeric matrix of read counts with rows=genes and columns=samples}

\item{remove_genes_ICB_proxies}{boolean variable to remove all those genes involved
in the computation of ICB proxy's of response}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
A list with the following elements:
\describe{
\item{scores}{pathway activity matrix with rows=samples and columns=pathways}
\item{transcripts_kept}{vector with available gene names}
\item{transcripts_left}{vector with missing gene names}
}
}
\description{
\code{compute_pathways_scores} infers pathway activity from raw counts RNA-seq data.
}
\examples{
# Example: Riaz
data("Riaz_data")
library("progeny")

# Computation of pathway scores
pathway_activity <- compute_pathways_scores(
  RNA_counts = Riaz_data$raw_counts_RNAseq,
  remove_genes_ICB_proxies = TRUE)
head(pathway_activity)
}
