% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_RIR.R
\name{compute_RIR}
\alias{compute_RIR}
\title{Compute Immune resistance program}
\usage{
compute_RIR(RNA_tpm)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=IRP
}
\description{
\code{compute_IRP} computes immune resistance program using the code provided by
the authors. (Jerby-Arnon et al., 2018).
TODOTODOoscar: out of sync name vs doc
}
\examples{
# Example: Riaz
data("Riaz_data")

RIR <- compute_RIR(RNA_tpm = Riaz_data$tpm_RNAseq)
head(RIR)
}
