% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_RIR.R
\name{compute_RIR}
\alias{compute_RIR}
\title{Compute Immune resistance program}
\usage{
compute_RIR(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{A numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=IRP
}
\description{
\code{compute_IRP} computes immune resistance program using the code provided by
the authors. (Jerby-Arnon et al., 2018).
TODOTODOoscar: out of sync name vs doc
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

# compute RIR signature from Jerby-Arnon et al., Cell, 2018
RIR <- compute_RIR(RNA_tpm = gene_tpm)
}
