% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Roh_IS.R
\name{compute_Roh_IS}
\alias{compute_Roh_IS}
\title{Compute Roh immune score}
\usage{
compute_Roh_IS(RNA_tpm)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=Roh immune score
}
\description{
\code{compute_Roh_IS} computes Roh immune score as the geometric-mean of immune
score genes (Roh et al., 2017).
}
\examples{
# Example: Riaz
data("Riaz_data")

Roh_IS <- compute_Roh_IS(RNA_tpm = Riaz_data$tpm_RNAseq)
head(Roh_IS)
}
