% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_signatures.R
\name{compute_Roh_IS}
\alias{compute_Roh_IS}
\title{Compute Roh immune score}
\usage{
compute_Roh_IS(matches, RNA.tpm)
}
\arguments{
\item{matches}{TODOTODO}

\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=Roh immune score
}
\description{
\code{compute_roh_IS} computes Roh immune score as the geometric-mean of immune score genes
(Roh et al., 2017).
}
\examples{
# TODOTODO
}
