% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Roh_IS.R
\name{compute_Roh_IS}
\alias{compute_Roh_IS}
\title{Compute Roh immune score (Roh_IS)}
\usage{
compute_Roh_IS(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{data.frame containing TPM values with HGNC symbols in rows and samples in columns.}

\item{verbose}{logical value indicating whether to display informative messages.}
}
\value{
A numeric matrix with samples in rows and Roh_IS score in a column.
}
\description{
This function computes Roh_IS score as the geometric-mean of its signature genes.
}
\examples{
# use example dataset from IMvigor210CoreBiologies package (Mariathasan et al., Nature, 2018)
data("dataset_mariathasan")
gene_tpm <- dataset_mariathasan@tpm

# compute Roh immune score (Roh et al., Sci. Transl. Med., 2017)
Roh_IS <- compute_Roh_IS(RNA_tpm = gene_tpm)
}
\references{
Roh, W., Chen, P.-L., Reuben, A., Spencer, C.N., Prieto, P.A., Miller, J.P., Gopalakrishnan, V.,
Wang, F., Cooper, Z.A., Reddy, S.M., et al. (2017). Integrated molecular analysis of tumor biopsies on sequential
CTLA-4 and PD-1 blockade reveals markers of response and resistance. Sci. Transl. Med. 9.
https://doi.org/10.1126/scitranslmed.aah3560.
}
