% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Tcell_inflamed.R
\name{compute_Tcell_inflamed}
\alias{compute_Tcell_inflamed}
\title{Compute T cell-inflamed signature score}
\usage{
compute_Tcell_inflamed(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=T cell-inflamed signature
score
}
\description{
\code{compute_ayersTcellInfl} computes T cell-inflamed signature score by
taking a weighted sum of the housekeeping normalized values of the T cell-inflamed
signature genes
TODOTODOoscar, async name of function vs doc
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

# compute T-cell inflamed signature (Ayers et al., JCI, 2017)
Tcell_inflamed <- compute_Tcell_inflamed(RNA_tpm = gene_tpm)
}
