% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_signatures.R
\name{compute_Tcell_inflamed}
\alias{compute_Tcell_inflamed}
\title{Compute T cell-inflamed signature score}
\usage{
compute_Tcell_inflamed(housekeeping, predictors, weights, RNA.tpm)
}
\arguments{
\item{housekeeping}{TODOTODO}

\item{predictors}{TODOTODO}

\item{weights}{TODOTODO}

\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=T cell-inflamed signature score
}
\description{
\code{compute_ayersTcellInfl} computes T cell-inflamed signature score by taking a weighted sum of
the housekeeping normalized values of the T cell-inflamed signature genes
}
\examples{
# TODOTODO
}
