% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Tcell_inflamed.R
\name{compute_Tcell_inflamed}
\alias{compute_Tcell_inflamed}
\title{Compute T cell-inflamed signature score}
\usage{
compute_Tcell_inflamed(RNA_tpm)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=T cell-inflamed signature
score
}
\description{
\code{compute_ayersTcellInfl} computes T cell-inflamed signature score by
taking a weighted sum of the housekeeping normalized values of the T cell-inflamed
signature genes
TODOTODOoscar, async name of function vs doc
}
\examples{
# Example: Riaz
data("Riaz_data")

Tcell_inflamed <- compute_Tcell_inflamed(RNA_tpm = Riaz_data$tpm_RNAseq)
head(Tcell_inflamed)
}
