% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_TF_activity.R
\name{compute_TF_activity}
\alias{compute_TF_activity}
\title{Compute transcription factor activity from gene expression using DoRothEA}
\usage{
compute_TF_activity(RNA_tpm, remove_genes_ICB_proxies = FALSE, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{A data.frame containing TPM values with HGNC symbols in rows and samples in columns.}

\item{remove_genes_ICB_proxies}{A logical value indicating whether to remove signature genes involved
in the derivation of hallmarks of immune response.}

\item{verbose}{A logical value indicating whether to display messages about the number of regulated
genes found in the gene expression data provided.}
}
\value{
a matrix with samples in rows and transcription factors in columns.
}
\description{
This function infers transcription factor activity from gene expression in TPM
from bulk RNA-seq data using DoRothEA method (Garcia-Alonso et al., Genome Res, 2019).
}
\examples{
# Example: Mariathasan cohort (Mariathasan et al., Nature, 2018)
if (!requireNamespace("BiocManager", quietly = TRUE))
 install.packages("BiocManager")

BiocManager::install(c("biomaRt",
 "circlize",
 "ComplexHeatmap",
 "corrplot",
 "DESeq2",
 "dplyr",
 "DT",
 "edgeR",
 "ggplot2",
 "limma",
 "lsmeans",
 "reshape2",
 "spatstat",
 "survival",
 "plyr"))

install.packages("Downloads/IMvigor210CoreBiologies_1.0.0.tar.gz", repos = NULL)
library(IMvigor210CoreBiologies)

data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

# Computation of TF activity
tf_activity <- compute_TF_activity(
  RNA_tpm = gene_tpm,
  remove_genes_ICB_proxies = FALSE)

head(tf_activity)
}
