% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_TF_activity.R
\name{compute_TF_activity}
\alias{compute_TF_activity}
\title{Compute transcription factors activity}
\usage{
compute_TF_activity(RNA_tpm, remove_genes_ICB_proxies = FALSE, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix of tpm values with rows=genes and columns=samples}

\item{remove_genes_ICB_proxies}{boolean variable to remove all those genes
involved in the computation of ICB proxy's of response}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
A list with the following elements:
\describe{
\item{scores}{TF activity matrix with rows=samples and columns=TFs}
\item{transcripts_kept}{vector with available gene names}
\item{transcripts_left}{vector with missing gene names}
}
}
\description{
\code{compute_TF_activity} infers transcription factor activity from tpm RNA-seq data.
}
\examples{
# Example: Riaz
data("Riaz_data")

# Computation of TF activity
tf_activity <- compute_TF_activity(
  RNA_tpm = Riaz_data$tpm_RNAseq,
  remove_genes_ICB_proxies = FALSE)

head(tf_activity)
}
