% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_TLS.R
\name{compute_TLS}
\alias{compute_TLS}
\title{Compute tertiary lymphoid structures signature}
\usage{
compute_TLS(RNA_tpm, verbose = TRUE)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}

\item{verbose}{A logical value indicating whether to display informative messages}
}
\value{
numeric matrix with rows=samples and columns=TLS signature
}
\description{
\code{compute_TLS} computes TLS signature as the geometric-mean of TLS signature
genes (Cabrita et al., 2020).
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_tpm <- mariathasan_data$tpm
rm(cds)

# compute tertiary lymphoid structures signature (Cabrita et al., Nature, 2020)
TLS <- compute_TLS(RNA_tpm = gene_tpm)
}
