% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_signatures.R
\name{compute_TLS}
\alias{compute_TLS}
\title{Computation of tertiary lymphoid structures signature (TLS) score}
\usage{
compute_TLS(matches, RNA_tpm, verbose)
}
\arguments{
\item{matches}{numeric vector indicating the index of signature genes in \code{RNA_tpm}.}

\item{RNA_tpm}{data.frame containing TPM values with HGNC symbols in rows and samples in columns.}

\item{verbose}{logical value indicating whether to display informative messages.}
}
\value{
A numeric matrix with samples in rows and TLS score in a column.
}
\description{
This function calculates TLS score as the geometric-mean of the expression of its signature genes.
}
\examples{
# use example dataset from IMvigor210CoreBiologies package (Mariathasan et al., Nature, 2018)
data("dataset_mariathasan")
gene_tpm <- dataset_mariathasan@tpm

# compute tertiary lymphoid structures signature (Cabrita et al., Nature, 2020)
TLS <- compute_TLS(RNA_tpm = gene_tpm)
}
\references{
Cabrita, R., Lauss, M., Sanna, A., Donia, M., Skaarup Larsen, M., Mitra, S., Johansson, I., Phung, B.,
Harbst, K., Vallon-Christersson, J., et al. (2020). Tertiary lymphoid structures improve immunotherapy and survival
in melanoma. Nature 577, 561–565.
}
