% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_TLS.R
\name{compute_TLS}
\alias{compute_TLS}
\title{Compute tertiary lymphoid structures signature}
\usage{
compute_TLS(RNA_tpm)
}
\arguments{
\item{RNA_tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=TLS signature
}
\description{
\code{compute_TLS} computes TLS signature as the geometric-mean of TLS signature
genes (Cabrita et al., 2020).
}
\examples{
# Example: Riaz
data("Riaz_data")

TLS <- compute_TLS(RNA_tpm= Riaz_data$tpm_RNAseq)
}
