% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easier-pkg.R
\docType{package}
\name{easier-pkg}
\alias{easier-pkg}
\title{Estimate Systems Immune Response}
\description{
pkg description
}
\references{
Lapuente-Santana, Óscar, Maisa van Genderen, Peter A. J. Hilbers, Francesca Finotello, and Federica Eduati.
2021. “Interpretable Systems Biomarkers Predict Response to Immune-Checkpoint Inhibitors.” Patterns, 100293.
https://doi.org/10.1016/j.patter.2021.100293.
}
