% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_biomarkers.R
\name{explore_biomarkers}
\alias{explore_biomarkers}
\title{Examination of possible relevant biomarkers in cancer dataset}
\usage{
explore_biomarkers(
  pathways = NULL,
  immunecells = NULL,
  tfs = NULL,
  lrpairs = NULL,
  ccpairs = NULL,
  cancer_type,
  real_patient_response,
  output_file_path,
  TMB_values,
  verbose = TRUE
)
}
\arguments{
\item{pathways}{A numeric matrix (rows = samples; columns = pathways).}

\item{immunecells}{A numeric matrix (rows = samples; columns = cell types).}

\item{tfs}{A numeric matrix (rows = samples; columns = transcription factors).}

\item{lrpairs}{A numeric matrix (rows = samples; columns = ligand-receptor pairs).}

\item{ccpairs}{A numeric matrix (rows = samples; columns = cell-cell pairs).}

\item{cancer_type}{A character string indicating which cancer-specific model should be used to compute the predictions.}

\item{real_patient_response}{A character vector with two factors (Non-responders = NR, Responders = R).}

\item{output_file_path}{A character string pointing to a directory to save the plots returned by the function.}

\item{verbose}{A logical flag indicating whether to display messages about the process.}
}
\value{
boxplot with features distribution
}
\description{
\code{explore_biomarkers} provides an overview of the mechanistic signatures features
distribution in the dataset (separating responders and non-responders if known)
together with information on the feature contribution to the model building.
}
\details{
Plot original features distribution together with the feature importance in
the model
}
\examples{
# Example: Mariathasan cohort (Mariathasan et al., Nature, 2018)
if (!requireNamespace("BiocManager", quietly = TRUE))
 install.packages("BiocManager")

BiocManager::install(c("biomaRt",
 "circlize",
 "ComplexHeatmap",
 "corrplot",
 "DESeq2",
 "dplyr",
 "DT",
 "edgeR",
 "ggplot2",
 "limma",
 "lsmeans",
 "reshape2",
 "spatstat",
 "survival",
 "plyr"))

install.packages("Downloads/IMvigor210CoreBiologies_1.0.0.tar.gz", repos = NULL)
library(IMvigor210CoreBiologies)

data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_count <- mariathasan_data$counts
gene_tpm <- mariathasan_data$tpm
rm(cds)

# Computation of cell fractions
cell_fractions <- compute_cell_fractions(RNA_tpm = gene_tpm)

# Computation of pathway scores
pathway_activity <- compute_pathways_scores(RNA_counts = gene_count,
remove_genes_ICB_proxies = TRUE)

# Computation of TF activity
tf_activity <- compute_TF_activity(RNA_tpm = gene_tpm,
remove_genes_ICB_proxies = FALSE)

# Computation of ligand-receptor pair weights
lrpair_weights <- compute_LR_pairs(RNA_tpm = gene_tpm,
remove_genes_ICB_proxies = FALSE,
cancer_type = "pancan")

# Computation of cell-cell interaction scores
ccpair_scores <- compute_CC_pairs_grouped(lrpairs = lrpair_weights,
cancer_type = "pancan")

# retrieve clinical response
clinical_data <- mariathasan_data$clinical
patient_response <- clinical_data[, "Best Confirmed Overall Response"]
patient_response <- gsub("CR","R", patient_response)
patient_response <- gsub("PD","NR", patient_response)
names(patient_response) <- rownames(clinical_data)
# retrieve TMB
TMB <- clinical_data[, "FMOne mutation burden per MB"]
names(TMB) <- rownames(clinical_data)

# Investigate possible biomarkers
explore_biomarkers(pathways = pathway_activity,
immunecells = cell_fractions,
lrpairs = lrpair_weights,
tfs = tf_activity,
ccpairs = ccpair_scores,
cancer_type = "BLCA",
real_patient_response = patient_response,
output_file_path = "../figures",
TMB_values = TMB)

}
