% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_biomarkers.R
\name{explore_biomarkers}
\alias{explore_biomarkers}
\title{Visualization of stunning biomarkers found in the dataset}
\usage{
explore_biomarkers(
  pathways = NULL,
  immunecells = NULL,
  tfs = NULL,
  lrpairs = NULL,
  ccpairs = NULL,
  cancer_type,
  real_patient_response,
  output_file_path,
  TMB_values,
  verbose = TRUE
)
}
\arguments{
\item{pathways}{numeric matrix with pathways activity (rows = samples; columns = pathways).}

\item{immunecells}{numeric matrix with immune cell quantification (rows = samples; columns = cell types).}

\item{tfs}{numeric matrix with transcription factors activity (rows = samples; columns = transcription factors).}

\item{lrpairs}{numeric matrix with ligand-receptor weights (rows = samples; columns = ligand-receptor pairs).}

\item{ccpairs}{numeric matrix with cell-cell scores (rows = samples; columns = cell-cell pairs).}

\item{cancer_type}{character string indicating which cancer-specific model should be used to compute the predictions.}

\item{real_patient_response}{character vector with two factors (Non-responders = NR, Responders = R).}

\item{output_file_path}{character string pointing to a directory to save the plots returned by the function.}

\item{TMB_values}{numeric vector containing patients' tumor mutational burden (TMB) values.}

\item{verbose}{logical flag indicating whether to display messages about the process.}
}
\value{
\itemize{
\item{Volcano plot displaying relevant biomarkers differentiating responders vs non-responders patients.}
\item{A combined plot for each type of quantitative descriptors, showing the original distribution of the features and the importance of these features for the trained models}
}
}
\description{
This function rovides an overview of relevant computed features (biomarkers), comparing responders and non-responders if known.
Information about the features contribution to the optimal models is also included.
}
\examples{
# use example dataset from IMvigor210CoreBiologies package (Mariathasan et al., Nature, 2018)
data("dataset_mariathasan")
gene_count <- dataset_mariathasan@counts
gene_tpm <- dataset_mariathasan@tpm

# Computation of cell fractions
cell_fractions <- compute_cell_fractions(RNA_tpm = gene_tpm)

# Computation of pathway scores
pathway_activity <- compute_pathway_activity(
  RNA_counts = gene_count,
  remove_sig_genes_immune_response = TRUE
)

# Computation of TF activity
tf_activity <- compute_TF_activity(
  RNA_tpm = gene_tpm
)

# Computation of ligand-receptor pair weights
lrpair_weights <- compute_LR_pairs(
  RNA_tpm = gene_tpm
  cancer_type = "pancan"
)

# Computation of cell-cell interaction scores
ccpair_scores <- compute_CC_pairs(
  lrpairs = lrpair_weights,
  cancer_type = "pancan"
)

# retrieve clinical response
patient_response <- dataset_mariathasan@response

# retrieve TMB
TMB <- dataset_mariathasan@TMB

# Investigate possible biomarkers
explore_biomarkers(
  pathways = pathway_activity,
  immunecells = cell_fractions,
  lrpairs = lrpair_weights,
  tfs = tf_activity,
  ccpairs = ccpair_scores,
  cancer_type = "BLCA",
  real_patient_response = patient_response,
  output_file_path = "../figures",
  TMB_values = TMB
)
}
