% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OE_bulk.R
\name{get_OE_bulk}
\alias{get_OE_bulk}
\title{Calculate overall expression (OE)}
\usage{
get_OE_bulk(r, gene_sign = NULL, num_rounds = 1000, full_flag = FALSE)
}
\arguments{
\item{r}{list TODOTODO - needs some more info?}

\item{gene_sign}{string}

\item{num_rounds}{integer}

\item{full_flag}{boolean}
}
\value{
Random score
}
\description{
\code{get_OE_bulk} obtained from literature to calculate Immune resistance program
(Jerby-Arnon et al., 2018)
}
\examples{
# TODOTODO
}
