% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_betas_boxplot.R
\name{plot_betas_boxplot}
\alias{plot_betas_boxplot}
\title{Betas Boxplot}
\usage{
plot_betas_boxplot(x, filename = NULL, main = NULL, ...)
}
\arguments{
\item{x}{numeric list with data for each factor}

\item{filename, }{file name to write the boxplot with png format}

\item{main}{optional, string with plot title}
}
\value{
Betas boxplot
}
\description{
Plot Betas boxplot
}
