% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_immune_response.R
\name{predict_immune_response}
\alias{predict_immune_response}
\title{Function to compute predicted immune response}
\usage{
predict_immune_response(
  pathways = NULL,
  immunecells = NULL,
  tfs = NULL,
  lrpairs = NULL,
  ccpairs = NULL,
  cancer_type,
  verbose = TRUE
)
}
\arguments{
\item{pathways}{A numeric matrix (rows = samples; columns = pathways).}

\item{immunecells}{A numeric matrix (rows = samples; columns = cell types).}

\item{tfs}{A numeric matrix (rows = samples; columns = transcription factors).}

\item{lrpairs}{A numeric matrix (rows = samples; columns = ligand-receptor pairs).}

\item{ccpairs}{A numeric matrix (rows = samples; columns = cell-cell pairs).}

\item{cancer_type}{A character string indicating which cancer-specific model should be used to compute the predictions.}

\item{verbose}{A logical flag indicating whether to display messages about the process.}
}
\value{
A list containing the predictions for each quantitative descriptor and for each task.
Given that the model training was repeated 100 times with randomized-cross validation, a set of 100 predictions is returned.
}
\description{
\code{predict_immune_response} predicts patients' immune response using regularized multi-task linear regression (RMTLR).
The predicted immune response is calculated based on the input features data and the parameters learned during model training.
}
\examples{
# Example: Mariathasan cohort (Mariathasan et al., Nature, 2018)
if (!requireNamespace("BiocManager", quietly = TRUE))
 install.packages("BiocManager")

BiocManager::install(c("biomaRt",
 "circlize",
 "ComplexHeatmap",
 "corrplot",
 "DESeq2",
 "dplyr",
 "DT",
 "edgeR",
 "ggplot2",
 "limma",
 "lsmeans",
 "reshape2",
 "spatstat",
 "survival",
 "plyr"))

install.packages("Downloads/IMvigor210CoreBiologies_1.0.0.tar.gz", repos = NULL)
library(IMvigor210CoreBiologies)

data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_count <- mariathasan_data$counts
gene_tpm <- mariathasan_data$tpm
rm(cds)

# Computation of cell fractions
cell_fractions <- compute_cell_fractions(RNA_tpm = gene_tpm)

# Computation of pathway scores
pathway_activity <- compute_pathways_scores(RNA_counts = gene_count,
remove_genes_ICB_proxies = TRUE)

# Computation of TF activity
tf_activity <- compute_TF_activity(RNA_tpm = gene_tpm,
remove_genes_ICB_proxies = FALSE)

# Computation of ligand-receptor pair weights
lrpair_weights <- compute_LR_pairs(RNA_tpm = gene_tpm,
remove_genes_ICB_proxies = FALSE,
cancer_type = "pancan")

# Computation of cell-cell interaction scores
ccpair_scores <- compute_CC_pairs_grouped(lrpairs = lrpair_weights,
cancer_type = "pancan")

# Predict patients' immune response
predictions_immune_response <- predict_immune_response(pathways = pathway_activity,
immunecells = cell_fractions,
tfs = tf_activity,
lrpairs = lrpair_weights,
ccpairs = ccpair_scores,
cancer_type = "BLCA")
}
