% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_immune_response.R
\name{predict_immune_response}
\alias{predict_immune_response}
\title{Immune response prediction}
\usage{
predict_immune_response(
  pathways = NULL,
  immunecells = NULL,
  tfs = NULL,
  lrpairs = NULL,
  ccpairs = NULL,
  cancertype
)
}
\arguments{
\item{pathways}{numeric matrix with data}

\item{immunecells}{numeric matrix with data}

\item{tfs}{numeric matrix with data}

\item{lrpairs}{numeric matrix with data}

\item{ccpairs}{numeric matrix with data}

\item{cancertype}{string character}
}
\value{
Predictions for each model building.
}
\description{
\code{predict_immune_response} predicts immune response using two algorithms:
regularized multi-task linear regression (RMTLR) and bayesian efficient multi-kernel algorithm (BEMKL). While
BEMKL can exploit information across different input and output datasets,
RMTLR can only do so for response variables.
Another advantage of BEMKL is missing data handling, which is not the case
for the other algorithm.
}
\details{
These algorithms use model parameters learned during training on different
types of data in order to compute the immune response.
}
\examples{
# Example: Riaz
data("Riaz_data")

# Computation of cell fractions
cell_fractions <- compute_cell_fractions(RNA_tpm = Riaz_data$tpm_RNAseq)

# Computation of pathway scores
pathway_activity <- compute_pathways_scores(RNA_counts = Riaz_data$raw_counts_RNAseq,
remove_genes_ICB_proxies = TRUE)

# Computation of TF activity
tf_activity <- compute_TF_activity(RNA_tpm = Riaz_data$tpm_RNAseq,
remove_genes_ICB_proxies = FALSE)

# Computation of ligand-receptor pair weights
lrpair_weights <- compute_LR_pairs(RNA_tpm = Riaz_data$tpm_RNAseq,
remove_genes_ICB_proxies = FALSE, cancertype = "pancan")

# Computation of cell-cell interaction scores
ccpair_scores <- compute_CC_pairs_grouped(lrpairs = lrpairs_weights$LRpairs,
cancertype = "pancan")

# Predict patients' immune response
predictions_immune_response <- predict_immune_response(pathways = pathways_activity,
immunecells = cell_fractions,
lrpairs = lrpairs_weights,
tfs = tf_activity,
ccpairs = ccpairs_scores,
cancertype = "SKCM")
}
