% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_with_rmtlr.R
\name{predict_with_rmtlr}
\alias{predict_with_rmtlr}
\title{Collects Regularized Multi-Task Linear Regression (RMTLR) cancer-specfic model predictions}
\usage{
predict_with_rmtlr(
  view_name,
  view_info,
  view_data,
  opt_model_cancer_view_spec,
  opt_xtrain_stats_cancer_view_spec,
  verbose = TRUE
)
}
\arguments{
\item{view_name}{character string containing the name of the input view.}

\item{view_info}{character string informing about the family of the input data.}

\item{view_data}{list containing the data for each input view.}

\item{opt_model_cancer_view_spec}{cancer-view-specific model feature parameters learned during training.}

\item{opt_xtrain_stats_cancer_view_spec}{cancer-view-specific features mean and standard deviation of the training set.}

\item{verbose}{logical flag indicating whether to display messages about the process.}
}
\value{
A lList of predictions matrices, one for each tasks (rows = samples; columns = [runs).
}
\description{
This function puts together predicted immune response by using a cancer-specifc model learned
from training with RMTLR algorithm.
}
\examples{
# Load exemplary dataset (Mariathasan et al., Nature, 2018) from ExperimentHub easierData.
# Original processed data is available from IMvigor210CoreBiologies package.
library("ExperimentHub")
eh <- ExperimentHub()
easierdata_eh <- query(eh, c("easierData"))
dataset_mariathasan <- easierdata_eh[["EH6677"]]
RNA_tpm <- dataset_mariathasan@assays@data@listData[["tpm"]]
cancer_type <- dataset_mariathasan@metadata$cancertype

# Computation of cell fractions
cell_fractions <- compute_cell_fractions(RNA_tpm = RNA_tpm)

view_name <- "immunecells"
view_info <- c(immunecells = "gaussian")
view_data <- list(immunecells = as.data.frame(cell_fractions))

# Predict using rmtlr
prediction_view <- predict_with_rmtlr(
  view_name = view_name,
  view_info = view_info,
  view_data = view_data,
  cancer_type = cancer_type
)
}
