% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_with_rmtlr.R
\name{predict_with_rmtlr}
\alias{predict_with_rmtlr}
\title{Collects Regularized Multi-Task Linear Regression (RMTLR) cancer-specfic model predictions}
\usage{
predict_with_rmtlr(
  view_name,
  view_info,
  view_data,
  cancer_type,
  verbose = TRUE
)
}
\arguments{
\item{view_name}{character string containing the name of the input view.}

\item{view_info}{character string informing about the family of the input data.}

\item{view_data}{list containing the data for each input view.}

\item{cancer_type}{character string indicating cancer type to specify cancer-specific optimization model to be used.}

\item{verbose}{logical flag indicating whether to display messages about the process.}
}
\value{
List of predictions matrices, one for each tasks (rows = samples; columns = runs).
}
\description{
Collects predicted immune response by using a cancer-specifc model learned from training with RMTLR algorithm.
}
\examples{
# use example dataset from IMvigor210CoreBiologies package (Mariathasan et al., Nature, 2018)
data("dataset_mariathasan")
gene_tpm <- dataset_mariathasan@tpm

# Computation of cell fractions
cell_fractions <- compute_cell_fractions(RNA_tpm = gene_tpm)

view_name <- "immunecells"
view_info <- c(immunecells = "gaussian")
view_data <- list(immunecells = as.data.frame(cell_fractions))

# Predict using rmtlr
prediction_view <- predict_with_rmtlr(
  view_name = view_name,
  view_info = view_info,
  view_data = view_data,
  cancer_type = "SKCM"
)
}
