% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reannotate_genes.R
\name{reannotate_genes}
\alias{reannotate_genes}
\title{Gene re-annotation using HGNC symbols}
\usage{
reannotate_genes(cur_genes)
}
\arguments{
\item{cur_genes}{character string containing gene HGNC symbols to be consider for re-annotation.}
}
\value{
A data.frame with the old gene HGNC symbol and the new corresponding gene HGNC symbol.
}
\description{
This function performs gene re-annotation using curated data by the HGNC.
}
\examples{
# Load exemplary dataset (Mariathasan et al., Nature, 2018) from easierData.
# Original processed data is available from IMvigor210CoreBiologies package.
library("easierData")
dataset_mariathasan <- easierData::get_Mariathasan2018_PDL1_treatment()
RNA_tpm <- dataset_mariathasan@assays@data@listData[["tpm"]]

genes_info <- reannotate_genes(cur_genes = rownames(RNA_tpm))
}
