% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reannotate_genes.R
\name{reannotate_genes}
\alias{reannotate_genes}
\title{Gene re-annotation using HGNC symbols}
\usage{
reannotate_genes(cur_genes)
}
\arguments{
\item{cur_genes}{A character string containing gene HGNC symbols to be consider for re-annotation.}
}
\value{
A data.frame with the old gene HGNC symbol and the new corresponding gene HGNC symbol.
}
\description{
This function performs gene re-annotation using curated data by the HGNC.
}
\examples{
# use example dataset from Mariathasan cohort (Mariathasan et al., Nature, 2018)
data(cds)
mariathasan_data <- preprocess_mariathasan(cds)
gene_count <- mariathasan_data$counts
rm(cds)

genes_info <- map_genes(cur_genes = rownames(gene_count))
}
