% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reannotate_genes.R
\name{reannotate_genes}
\alias{reannotate_genes}
\title{Gene re-annotation using HGNC symbols}
\usage{
reannotate_genes(cur_genes)
}
\arguments{
\item{cur_genes}{character string containing gene HGNC symbols to be consider for re-annotation.}
}
\value{
Data.frame with the old gene HGNC symbol and the new corresponding gene HGNC symbol.
}
\description{
Performs gene re-annotation using curated data by the HGNC.
}
\examples{
# use example dataset from IMvigor210CoreBiologies package (Mariathasan et al., Nature, 2018)
data("dataset_mariathasan")
gene_tpm <- dataset_mariathasan@tpm

genes_info <- reannotate_genes(cur_genes = rownames(gene_tpm))
}
