% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmtlr_test.R
\name{rmtlr_test}
\alias{rmtlr_test}
\title{Regularized Multi-Task Linear Regression (RMTLR) model predictions}
\usage{
rmtlr_test(x_test, coef_matrix)
}
\arguments{
\item{x_test}{numeric matrix containing features values (rows = samples; columns = features).}

\item{coef_matrix}{numeric matrix containing the parameters values derived from model training
(rows = features; columns = tasks).}
}
\value{
Numeric matrix of predicted values (rows = samples; columns = tasks).
}
\description{
Computes the predictions from the results of RMTLR models optimization
}
\examples{
# use example dataset from IMvigor210CoreBiologies package (Mariathasan et al., Nature, 2018)
data("dataset_mariathasan")
gene_tpm <- dataset_mariathasan@tpm

# Computation of cell fractions (Finotello et al., Genome Med, 2019)
cell_fractions <- compute_cell_fractions(RNA_tpm = gene_tpm)

# Parameters values should be defined with features as rows and tasks as columns
estimated_parameters <- matrix(rnorm(n = (ncol(cell_fractions) + 1) * 10),
  nrow = ncol(cell_fractions) + 1, ncol = 10
)
rownames(estimated_parameters) <- c("(Intercept)", colnames(cell_fractions))
colnames(estimated_parameters) <- c(
  "CYT", "Ock_IS", "Roh_IS", "chemokines",
  "Davoli_IS", "IFNy", "Ayers_expIS", "Tcell_inflamed", "RIR", "TLS"
)

# Compute predictions using parameters values
pred_test <- rmtlr_test(
  x_test = cell_fractions,
  coef_matrix = estimated_parameters
)
}
