% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmtlr_test.R
\name{rmtlr_test}
\alias{rmtlr_test}
\title{Regularized multi-task linear regression algorithm implements the model
training coefficients and hyperparameters on the test set.}
\usage{
rmtlr_test(x_test, coef_matrix)
}
\arguments{
\item{x_test}{A matrix with data required to perform predictions.}

\item{coef_matrix}{A coefficient matrix obtained during training the model
(rows = features and columns = tasks).}
}
\value{
A matrix with predictions for each sample and task.
}
\description{
\code{rmtlr_test} implements regularized multi-task linear regression on the test data
}
\examples{
# TODOTODO
}
